% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_funcoup_functions.R
\name{add_annotation_funcoup}
\alias{add_annotation_funcoup}
\title{add_annotation_funcoup ()}
\usage{
add_annotation_funcoup(ppi_funcoup, anno_df, species)
}
\arguments{
\item{ppi_funcoup}{variable defined by ppis_funcoup in get_networkdata_funcoup()}

\item{anno_df}{annotation dataframe (for corresponding species in funcoup)}

\item{species}{from which species does the data come from}
}
\value{
ppi_funcoup with annotation columns for each interactor (for corresponding species in funcoup)
}
\description{
add_annotation_funcoup ()
}
\examples{
\dontrun{
db_funcoup_df <- get_networkdata_funcoup( species = "H.sapiens",
                                          version = "5.0",
                                          cache = TRUE,
                                          get_annotation = FALSE,
                                          add_annotation = FALSE
                                         )

db_funcoup_anno_df <- get_annotation_funcoup(ppi_funcoup = db_funcoup_df,
                                       species = "H.sapiens"
                                       )

db_funcoup_ppi_anno_df <- add_annotation_funcoup(  ppi_funcoup = db_funcoup_df,
                                                   anno_df = db_funcoup_anno_df,
                                                   species = "H.sapiens"
                                                   )

}
}
