% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_genemania_functions.R
\name{add_annotation_genemania}
\alias{add_annotation_genemania}
\title{add_annotation_genemania ()}
\usage{
add_annotation_genemania(ppi_genemania, anno_df, species)
}
\arguments{
\item{ppi_genemania}{variable defined by ppis_genemania in get_networkdata_genemania()}

\item{anno_df}{annotation dataframe (for corresponding species in genemania)}

\item{species}{from which species does the data come from}
}
\value{
ppi_genemania with annotation columns for each interactor (for corresponding species in genemania)
}
\description{
add_annotation_genemania ()
}
\examples{
\dontrun{
db_genemania_df <- get_networkdata_genemania(species = "Homo_sapiens",
                                             version = "current",
                                             get_annotation = FALSE,
                                             add_annotation = FALSE
                                             )

db_genemania_anno_df <- get_annotation_genemania( ppi_genemania = db_genemania_df,
                                                  species = "Homo_sapiens"
                                                )

db_genemania_ppi_anno_df <- add_annotation_genemania( ppi_genemania = db_genemania_df,
                                                      anno_df = db_genemania_anno_df,
                                                      species = "Homo_sapiens"
                                                    )

}
}
