% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_hint_functions.R
\name{add_annotation_hint}
\alias{add_annotation_hint}
\title{add_annotation_hint ()}
\usage{
add_annotation_hint(ppi_hint, anno_df, species)
}
\arguments{
\item{ppi_hint}{variable defined by ppis_hint in get_networkdata_hint()}

\item{anno_df}{annotation dataframe (for corresponding species in hint)}

\item{species}{from which species does the data come from}
}
\value{
ppi_hint with annotation columns for each interactor (for corresponding species in hint)
}
\description{
add_annotation_hint ()
}
\examples{
\dontrun{
db_hint_df <- get_networkdata_hint(species = "HomoSapiens",
                                   version = "2024-06",
                                   type = "binary",
                                   get_annotation = FALSE,
                                   add_annotation = FALSE
)

db_hint_anno_df <- get_annotation_hint(ppi_hint = db_hint_df,
                                       species = "HomoSapiens"
                                       )

db_hint_ppi_anno_df <- add_annotation_hint(ppi_hint = db_hint_df,
                                           anno_df = db_hint_anno_df,
                                           species = "HomoSapiens"
                                           )

}
}
