% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_iid_functions.R
\name{add_annotation_iid}
\alias{add_annotation_iid}
\title{add_annotation_iid ()}
\usage{
add_annotation_iid(ppi_iid, anno_df, species)
}
\arguments{
\item{ppi_iid}{variable defined by ppis_iid in get_networkdata_iid()}

\item{anno_df}{annotation dataframe (for corresponding species in iid)}

\item{species}{from which species does the data come from}
}
\value{
ppi_iid with annotation columns for each interactor (for corresponding species in iid)
}
\description{
add_annotation_iid ()
}
\examples{
\dontrun{
db_iid_df <- get_networkdata_iid(species = "human",
                                 version = "2021-05",
                                 get_annotation = FALSE,
                                 add_annotation = FALSE
                                )

db_iid_anno_df <- get_annotation_iid( ppi_iid = db_iid_df,
                                      species = "human"
                                      )

db_iid_ppi_anno_df <- add_annotation_iid(ppi_iid = db_iid_df,
                                         anno_df = db_iid_anno_df,
                                         species = "human"
                                         )

}
}
