% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_innatedb_functions.R
\name{add_annotation_innatedb}
\alias{add_annotation_innatedb}
\title{add_annotation_innatedb ()}
\usage{
add_annotation_innatedb(ppi_innatedb, anno_df, species)
}
\arguments{
\item{ppi_innatedb}{variable defined by ppis_innatedb in get_networkdata_innatedb()}

\item{anno_df}{annotation dataframe (for corresponding species in innatedb)}

\item{species}{from which species does the data come from}
}
\value{
ppi_innatedb with annotation columns for each interactor (for corresponding species in innatedb)
}
\description{
add_annotation_innatedb ()
}
\examples{
\dontrun{
db_innatedb_df <- get_networkdata_innatedb( species = "taxid:9606(Human)",
                                            version = "5.4",
                                            cache = TRUE,
                                            get_annotation = FALSE,
                                            add_annotation = FALSE
                                            )

db_innatedb_anno_df <- get_annotation_innatedb( ppi_innatedb = db_innatedb_df,
                                                species = "taxid:9606(Human)"
                                                )

db_innatedb_ppi_anno_df <- add_annotation_innatedb( ppi_innatedb = db_innatedb_df,
                                                    anno_df = db_innatedb_anno_df,
                                                    species = "taxid:9606(Human)"
                                                  )

}
}
