% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_intact_functions.R
\name{add_annotation_intact}
\alias{add_annotation_intact}
\title{add_annotation_intact ()}
\usage{
add_annotation_intact(ppi_intact, anno_df, species)
}
\arguments{
\item{ppi_intact}{variable defined by ppis_intact in get_networkdata_intact()}

\item{anno_df}{annotation dataframe (for corresponding species in intact)}

\item{species}{from which species does the data come from}
}
\value{
ppi_intact with annotation columns for each interactor (for corresponding species in intact)
}
\description{
add_annotation_intact ()
}
\examples{
\dontrun{
db_intact_df <- get_networkdata_intact( species = "taxid:9606(human)|taxid:9606(Homo sapiens)",
                                        version = "current",
                                        cache = TRUE,
                                        get_annotation = FALSE,
                                        add_annotation = FALSE
                                        )

db_intact_anno_df <- get_annotation_intact( ppi_intact = db_intact_df,
                                            species = "taxid:9606(human)|taxid:9606(Homo sapiens)"
                                            )

db_intact_ppi_anno_df <- add_annotation_intact( ppi_intact = db_intact_df,
                                                anno_df = db_intact_anno_df,
                                                species = "taxid:9606(human)|taxid:9606(Homo sapiens)"
                                                 )

}
}
