% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_irefindex_functions.R
\name{add_annotation_irefindex}
\alias{add_annotation_irefindex}
\title{add_annotation_irefindex ()}
\usage{
add_annotation_irefindex(ppi_irefindex, anno_df, species)
}
\arguments{
\item{ppi_irefindex}{variable defined by ppis_irefindex in get_networkdata_irefindex()}

\item{anno_df}{annotation dataframe (for corresponding species in irefindex)}

\item{species}{from which species does the data come from}
}
\value{
ppi_irefindex with annotation columns for each interactor (for corresponding species in irefindex)
}
\description{
add_annotation_irefindex ()
}
\examples{
\dontrun{
db_irefindex_df <- get_networkdata_irefindex(
  species = "Homo sapiens",
  version = "08-28-2023",
  get_annotation = FALSE,
  add_annotation = FALSE
)

db_irefindex_anno_df <- get_annotation_irefindex(
  ppi_irefindex = db_irefindex_df,
  species = "Homo sapiens"
)

db_irefindex_ppi_anno_df <- add_annotation_irefindex(
  ppi_irefindex = db_irefindex_df,
  anno_df = db_irefindex_anno_df,
  species = "Homo sapiens"
)
}
}
