% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_pathwaycommons_functions.R
\name{add_annotation_pathwaycommons}
\alias{add_annotation_pathwaycommons}
\title{add_annotation_pathwaycommons ()}
\usage{
add_annotation_pathwaycommons(ppi_pathwaycommons, anno_df, species)
}
\arguments{
\item{ppi_pathwaycommons}{variable defined by ppis_pathwaycommons in get_networkdata_pathwaycommons()}

\item{anno_df}{annotation dataframe (for corresponding species in pathwaycommons)}

\item{species}{from which species does the data come from}
}
\value{
ppi_pathwaycommons with annotation columns for each interactor (for corresponding species in pathwaycommons)
}
\description{
add_annotation_pathwaycommons ()
}
\examples{
\dontrun{
db_pathwaycommons_df <- get_networkdata_pathwaycommons( species = "human",
                                                        version = "v12",
                                                        cache = TRUE,
                                                        get_annotation = FALSE,
                                                        add_annotation = FALSE
                                                        )

db_pathwaycommons_anno_df <- get_annotation_pathwaycommons( ppi_pathwaycommons = db_pathwaycommons_df,
                                                            species = "human")

db_pathwaycommons_ppi_anno_df <- add_annotation_pathwaycommons( ppi_pathwaycommons = db_pathwaycommons_df,
                                                                anno_df = db_pathwaycommons_anno_df,
                                                                species = "human"
                                                              )

}
}
