% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_reactome_functions.R
\name{add_annotation_reactome}
\alias{add_annotation_reactome}
\title{add_annotation_reactome ()}
\usage{
add_annotation_reactome(ppi_reactome, anno_df, species)
}
\arguments{
\item{ppi_reactome}{variable defined by ppis_reactome in get_networkdata_reactome()}

\item{anno_df}{annotation dataframe (for corresponding species in reactome)}

\item{species}{from which species does the data come from}
}
\value{
ppi_reactome with annotation columns for each interactor (for corresponding species in reactome)
}
\description{
add_annotation_reactome ()
}
\examples{
\dontrun{
db_reactome_df <- get_networkdata_reactome(
  species = "taxid:9606(Homo sapiens)",
  version = "current",
  cache = TRUE,
  get_annotation = FALSE,
  add_annotation = FALSE
)

db_reactome_anno_df <- get_annotation_reactome(
  ppi_reactome = db_reactome_df,
  species = "taxid:9606(Homo sapiens)"
)

db_reactome_ppi_anno_df <- add_annotation_reactome(
  ppi_reactome = db_reactome_df,
  anno_df = db_reactome_anno_df,
  species = "taxid:9606(Homo sapiens)"
)
}
}
