% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-1_db_stringdb.R
\name{add_annotation_stringdb}
\alias{add_annotation_stringdb}
\title{add_annotation_stringdb ()}
\usage{
add_annotation_stringdb(ppi_stringdb, anno_df, species)
}
\arguments{
\item{ppi_stringdb}{variable defined by ppis_stringdb in get_networkdata_stringdb()}

\item{anno_df}{annotation dataframe (for corresponding species in stringdb)}

\item{species}{from which species does the data come from}
}
\value{
ppi_stringdb with annotation columns for each interactor (for corresponding species in stringdb)
}
\description{
add_annotation_stringdb ()
}
\examples{
\dontrun{

db_stringdb_df <- get_networkdata_stringdb(species = "Homo sapiens",
                                           version = "12.0",
                                           cache = TRUE,
                                           get_annotation = FALSE,
                                           add_annotation = FALSE
                                          )

db_stringdb_anno_df <- get_annotation_stringdb(species = "Homo sapiens",
                                            version = "12.0",
                                            cache = TRUE
                                            )

db_stringdb_ppi_anno_df <- add_annotation_stringdb(ppi_stringdb = db_stringdb_df,
                                                   anno_df = db_stringdb_anno_df,
                                                   species = "Homo sapiens"
                                                   )

}
}
