% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_utils.R
\name{add_info_from_dataframe_to_graph}
\alias{add_info_from_dataframe_to_graph}
\title{add_info_from_dataframe_to_graph()}
\usage{
add_info_from_dataframe_to_graph(graph_data, g)
}
\arguments{
\item{graph_data}{dataframe of ppi data from corresponding database}

\item{g}{graph object created with build_graph()}
}
\value{
g
}
\description{
add_info_from_dataframe_to_graph()
}
\examples{

db_matrixdb_df <- get_networkdata_matrixdb(
  species = "human",
  type = "CORE"
)

igraph_object_matrixdb <- build_graph_matrixdb(graph_data = db_matrixdb_df,
                                  output_format = "igraph",
                                  min_score_threshold = NULL)

igraph_object_matrixdb_info_added <- add_info_from_dataframe_to_graph(graph_data = db_matrixdb_df,
                                                                     g = igraph_object_matrixdb)

igraph_object_matrixdb_info_added


}
