% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-3_db_funcoup.R
\name{annotation_funcoup}
\alias{annotation_funcoup}
\title{annotation_funcoup ()}
\usage{
annotation_funcoup(ppis_funcoup, species, version)
}
\arguments{
\item{ppis_funcoup}{variable defined by ppis_funcoup in get_networkdata_funcoup()}

\item{species}{from which species does the data come from c( "A.thaliana", "B.subtilis", "B.taurus", "C.elegans","C.familiaris", "C.intestinalis", "D.melanogatser", "D.rerio", "E.coli", "G.gallus", "H.sapiens", "M.jannaschii", "M.musculus", "O.sativa", "P.falciparum", "R.norvegicus", "S.cerevisae", "S.pombe", "S.scrofa", "S.solfataricus")}

\item{version}{version of the data files in funcoup}
}
\value{
ppis_funcoup_annotated
}
\description{
annotation_funcoup ()
}
\examples{

annotation_funcoup(ppis_funcoup, species = "H.sapiens", version = "5.0")

}
