% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-7_db_genemania.R
\name{annotation_genemania}
\alias{annotation_genemania}
\title{annotation_genemania ()}
\usage{
annotation_genemania(ppi_genemania, species, version)
}
\arguments{
\item{ppi_genemania}{variable defined by ppis_genemania in get_networkdata_genemania()}

\item{species}{from which species does the data come from}

\item{version}{version of the data files in genemania}
}
\value{
ppis_genemania_annotated
}
\description{
annotation_genemania ()
}
\examples{

# annotation_genemania(ppi_genemania, species = "HomoSapiens", version = "2024-06", type = "binary")
#TODO: what can I do here as ppi_genemania is not defined in annotation_genemania()?
}
