% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-2_db_hint.R
\name{annotation_hint}
\alias{annotation_hint}
\title{annotation_hint ()}
\usage{
annotation_hint(ppi_hint, species = "HomoSapiens", version, type)
}
\arguments{
\item{type}{}
}
\value{
ppis_hint_annotated
}
\description{
annotation_hint ()
}
\examples{

annotation_hint(species = "HomoSapiens", version = "2024-06", type = "binary")

}
