% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-4_db_iid.R
\name{annotation_iid}
\alias{annotation_iid}
\title{annotation_iid ()}
\usage{
annotation_iid(ppi_iid, species, version)
}
\arguments{
\item{ppi_iid}{variable defined by ppis_iid in get_networkdata_iid()}

\item{species}{from which species does the data come from}

\item{version}{version of the data files in iid}

\item{type}{different interaction files provided by iid (all high-quality)}
}
\value{
ppis_iid_annotated
}
\description{
annotation_iid ()
}
\examples{

# annotation_iid(ppi_iid, species = "HomoSapiens", version = "2024-06", type = "binary")
#TODO: what can I do here as ppi_funcoup is not defined in annotation_iid()?
}
