% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-13_db_innatedb.R
\name{annotation_innatedb}
\alias{annotation_innatedb}
\title{annotation_innatedb ()}
\usage{
annotation_innatedb(ppi_innatedb, species, version)
}
\arguments{
\item{ppi_innatedb}{variable defined by ppis_innatedb in get_networkdata_innatedb()}

\item{species}{from which species does the data come from}

\item{version}{version of the data files in innatedb}
}
\value{
ppis_innatedb_annotated
}
\description{
annotation_innatedb ()
}
\examples{
\dontrun{
annotation_innatedb <- annotation_innatedb(ppi_innatedb,
                                           species = "taxid:9606(Homo sapiens)",
                                           version = "current")
annotation_innatedb
}
}
