% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-15_db_intact.R
\name{annotation_intact}
\alias{annotation_intact}
\title{annotation_intact ()}
\usage{
annotation_intact(ppi_intact, species, version)
}
\arguments{
\item{ppi_intact}{variable defined by ppis_intact in get_networkdata_intact()}

\item{species}{from which species does the data come from}

\item{version}{version of the data files in intact}
}
\value{
ppis_intact_annotated
}
\description{
annotation_intact ()
}
\examples{
#\donttest{
#annotation_intact <- annotation_intact(ppi_intact, species = "taxid:9606(Homo sapiens)", version = "current")
#annotation_intact
#}
}
