% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-5_db_irefindex.R
\name{annotation_irefindex}
\alias{annotation_irefindex}
\title{annotation_irefindex ()}
\usage{
annotation_irefindex(ppi_irefindex, species, version)
}
\arguments{
\item{ppi_irefindex}{variable defined by ppis_irefindex in get_networkdata_irefindex()}

\item{species}{from which species does the data come from}

\item{version}{version of the data files in irefindex}
}
\value{
ppis_irefindex_annotated
}
\description{
annotation_irefindex ()
}
\examples{
\dontrun{
annotation_irefindex <- annotation_irefindex(ppi_irefindex,
                                             species = "Homo sapiens",
                                             version = "08-28-2023")
annotation_irefindex
}
}
