% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-6_db_mint.R
\name{annotation_mint}
\alias{annotation_mint}
\title{annotation_mint ()}
\usage{
annotation_mint(ppi_mint, species, version)
}
\arguments{
\item{ppi_mint}{variable defined by ppis_mint in get_networkdata_mint()}

\item{species}{from which species does the data come from}

\item{version}{version of the data files in mint}

\item{type}{different interaction files provided by mint (all high-quality)}
}
\value{
ppis_mint_annotated
}
\description{
annotation_mint ()
}
\examples{

# annotation_mint(ppi_mint, species = "HomoSapiens", version = "2024-06", type = "binary")
#TODO: what can I do here as ppi_funcoup is not defined in annotation_mint()?
}
