% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-11_db_pathwaycommons.R
\name{annotation_pathwaycommons}
\alias{annotation_pathwaycommons}
\title{annotation_pathwaycommons ()}
\usage{
annotation_pathwaycommons(ppi_pathwaycommons, species, version)
}
\arguments{
\item{ppi_pathwaycommons}{variable defined by ppis_pathwaycommons in get_networkdata_pathwaycommons()}

\item{species}{from which species does the data come from}

\item{version}{version of the data files in pathwaycommons}
}
\value{
ppis_pathwaycommons_annotated
}
\description{
annotation_pathwaycommons ()
}
\examples{
#\donttest{
#annotation_pathwaycommons(ppi_pathwaycommons, species = "human", version = "v12")
#}
}
