% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-12_db_reactome.R
\name{annotation_reactome}
\alias{annotation_reactome}
\title{annotation_reactome ()}
\usage{
annotation_reactome(ppi_reactome, species, version)
}
\arguments{
\item{ppi_reactome}{variable defined by ppis_reactome in get_networkdata_reactome()}

\item{species}{from which species does the data come from}

\item{version}{version of the data files in reactome}
}
\value{
ppis_reactome_annotated
}
\description{
annotation_reactome ()
}
\examples{
#\donttest{
#annotation_reactome(ppi_reactome, species = "taxid:9606(Homo sapiens)", version = "current")
#}
}
