% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-1_db_stringdb.R
\name{annotation_stringdb}
\alias{annotation_stringdb}
\title{annotation_stringdb ()}
\usage{
annotation_stringdb(ppi_stringdb, species, version)
}
\arguments{
\item{ppi_stringdb}{variable defined by ppis_stringdb in get_networkdata_stringdb()}

\item{species}{from which species does the data come from}

\item{version}{version of the data files in stringdb}
}
\value{
ppis_stringdb_annotated
}
\description{
annotation_stringdb ()
}
\examples{
#\dontrun{
#annotation_stringdb <- annotation_stringdb(ppi_stringdb, species = "taxid:9606(Homo sapiens)", version = "current")
#annotation_stringdb
#}
}
