% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_utils.R
\name{build_graph}
\alias{build_graph}
\title{Title}
\usage{
build_graph(
  graph_data,
  data_source = c("stringdb", "hint", "funcoup", "iid", "irefindex", "mint", "genemania",
    "reactome", "cpdb", "huri", "matrixdb", "pathwaycommons", "innatedb", "biogrid",
    "corum", "intact"),
  output_format = c("igraph", "grpahnel", "sif"),
  min_score_threshold,
  subset_nodes
)
}
\arguments{
\item{graph_data}{data describing the nodes and edges of the network}

\item{data_source}{database (e.g.: "stringdb")}

\item{output_format}{vector that specifies possible output formats ("igraph", "grpahnel", "sif")}

\item{min_score_threshold}{optional threshold to filter edges based on their combination score}

\item{subset_nodes}{optional vector of nodes to extract only a part of the network}
}
\value{
my_graph object
}
\description{
Title
}
\examples{

TODO
}
