% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_biogrid_functions.R
\name{build_graph_biogrid}
\alias{build_graph_biogrid}
\title{build_graph_biogrid()}
\usage{
build_graph_biogrid(
  graph_data,
  output_format = "igraph",
  min_score_threshold = NULL
)
}
\arguments{
\item{graph_data}{ppi data from biogrid}

\item{output_format}{selection of different graph functions that can be used}

\item{min_score_threshold}{select ppis that are "confident" depending on the scoretype/value}
}
\value{
An \code{igraph} graph object
}
\description{
build_graph_biogrid()
}
\examples{
\dontrun{

db_biogrid_df <- get_networkdata_biogrid(
  species = "9606",
  version = "4.4.238"
)

db_biogrid_graph <- build_graph_biogrid(
  graph_data = db_biogrid_df,
  output_format = "igraph"
)
db_biogrid_graph # list of 20124
}

}
