% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_genemania_functions.R
\name{build_graph_genemania}
\alias{build_graph_genemania}
\title{build_graph_genemania()}
\usage{
build_graph_genemania(
  graph_data,
  output_format = "igraph",
  min_score_threshold = NULL
)
}
\arguments{
\item{graph_data}{ppi data from genemania}

\item{output_format}{selection of different graph functions that can be used}

\item{min_score_threshold}{select ppis that are "confident" depending on the scoretype/value}
}
\value{
my_graph
}
\description{
build_graph_genemania()
}
\examples{
\dontrun{

db_genemania_df <- get_networkdata_genemania(
  species = "Homo_sapiens",
  version = "current"
)

db_genemania_graph <- build_graph_genemania(graph_data = db_genemania_df,
                                        output_format = "igraph",
                                        min_score_threshold = "0.00005")
db_genemania_graph #list of 18644
}


}
