% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_hint_functions.R
\name{build_graph_hint}
\alias{build_graph_hint}
\title{build_graph_hint()}
\usage{
build_graph_hint(
  graph_data,
  output_format = "igraph",
  min_score_threshold = NULL
)
}
\arguments{
\item{graph_data}{ppi data from HINT}

\item{output_format}{selection of different graph functions that can be used}

\item{min_score_threshold}{select ppis that are "confident" depending on the scoretype/value}
}
\value{
my_graph
}
\description{
build_graph_hint()
}
\examples{
\dontrun{
db_hint_df <- get_networkdata_hint(
  species = "HomoSapiens",
  version = "2024-06",
  type = "binary",
  cache = TRUE,
  get_annotation = TRUE,
  add_annotation = TRUE
)

db_hint_graph <- build_graph_hint(
  graph_data = db_hint_df,
  output_format = "igraph",
  min_score_threshold = NULL
)
db_hint_graph # list of ...TODO
}

}
