% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-9_db_huri.R
\name{build_graph_huri}
\alias{build_graph_huri}
\title{build_graph_huri()}
\usage{
build_graph_huri(graph_data, output_format = "igraph")
}
\arguments{
\item{graph_data}{ppi data from huri}

\item{output_format}{selection of different graph functions that can be used}

\item{min_score_treshold}{select ppis that are "confident" depending on the scoretype/value}
}
\description{
build_graph_huri()
}
\examples{
\dontrun{

db_huri_df <- get_networkdata_huri(
  species = "human",
  type = "HI-union"
)

db_huri_graph <- build_graph_huri(graph_data = db_huri_df,
                                        output_format = "igraph")
db_huri_graph #list of 9024
}


}
