% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-13_db_innatedb.R
\name{build_graph_innatedb}
\alias{build_graph_innatedb}
\title{build_graph_innatedb()}
\usage{
build_graph_innatedb(
  graph_data,
  output_format = "igraph",
  min_score_threshold = NULL
)
}
\arguments{
\item{graph_data}{ppi data from innatedb}

\item{output_format}{selection of different graph functions that can be used}

\item{min_score_treshold}{select ppis that are "confident": lpr score (lowest PMID re-use)}
}
\description{
build_graph_innatedb()
}
\examples{
\dontrun{

db_innatedb_df <- get_networkdata_innatedb(species = "taxid:9606(Human)",
                                           version = "5.4"
                                           )

db_innatedb_graph <- build_graph_innatedb(graph_data = db_innatedb_df,
                                          output_format = "igraph",
                                          min_score_threshold = "20" )
db_innatedb_graph #list of 690 (score = 10) 1835 (score = 20)
}


}
