% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-10_db_matrixdb.R
\name{build_graph_matrixdb}
\alias{build_graph_matrixdb}
\title{build_graph_matrixdb()}
\usage{
build_graph_matrixdb(graph_data, output_format = "igraph")
}
\arguments{
\item{graph_data}{ppi data from matrixdb}

\item{output_format}{selection of different graph functions that can be used}

\item{min_score_treshold}{select ppis that are "confident" depending on the scoretype/value}
}
\description{
build_graph_matrixdb()
}
\examples{
\dontrun{

db_matrixdb_df <- get_networkdata_matrixdb(
  species = "human",
  type = "CORE"
)

db_matrixdb_graph <- build_graph_matrixdb(graph_data = db_matrixdb_df,
                                        output_format = "igraph")
db_matrixdb_graph #list of 235
}


}
