% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-11_db_pathwaycommons.R
\name{build_graph_pathwaycommons}
\alias{build_graph_pathwaycommons}
\title{build_graph_pathwaycommons()}
\usage{
build_graph_pathwaycommons(
  graph_data,
  output_format = "igraph",
  min_score_threshold = NULL
)
}
\arguments{
\item{graph_data}{ppi data from pathwaycommons}

\item{output_format}{selection of different graph functions that can be used}

\item{min_score_threshold}{select ppis that are "confident" depending on the scoretype/value}
}
\description{
build_graph_pathwaycommons()
}
\examples{
\dontrun{

db_pathwaycommons_df <- get_networkdata_pathwaycommons(
  species = "human",
  version = "v12"
)

db_pathwaycommons_graph <- build_graph_pathwaycommons(graph_data = db_pathwaycommons_df,
                                        output_format = "igraph")
db_pathwaycommons_graph #list of 60434
}


}
