% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-12_db_reactome.R
\name{build_graph_reactome}
\alias{build_graph_reactome}
\title{build_graph_reactome()}
\usage{
build_graph_reactome(
  graph_data,
  output_format = "igraph",
  min_score_threshold = NULL
)
}
\arguments{
\item{graph_data}{ppi data from reactome}

\item{output_format}{selection of different graph functions that can be used}

\item{min_score_threshold}{select ppis that are "confident" depending on the scoretype/value}
}
\description{
build_graph_reactome()
}
\examples{
\dontrun{

db_reactome_df <- get_networkdata_reactome(species = "taxid:9606(Homo sapiens)",
                                           version = "current"
                                           )

db_reactome_graph <- build_graph_reactome(graph_data = db_reactome_df,
                                        output_format = "igraph",
                                        min_score_threshold = "0.5")
db_reactome_graph #list of 5416
}


}
