% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-1_db_stringdb.R
\name{build_graph_stringdb}
\alias{build_graph_stringdb}
\title{build_graph_stringdb()}
\usage{
build_graph_stringdb(
  graph_data,
  output_format = c("igraph", "graphnel", "sif"),
  min_score_threshold = NULL,
  subset_nodes = NULL
)
}
\arguments{
\item{graph_data}{data describing the nodes and edges of the network}

\item{output_format}{vector that specifies possible output formats ("igraph", "grpahnel", "sif")}

\item{min_score_threshold}{optional threshold to filter edges based on their combination score}

\item{subset_nodes}{optional vector of nodes to extract only a part of the network}
}
\value{
my_graph object
}
\description{
build_graph_stringdb()
}
\examples{

network_all <- build_graph_stringdb(graph_data = graph_data,
                                    output_format = "igraph",
                                    min_score_threshold = 600)
TODO
}
