% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-1_db_stringdb.R
\name{create_annotation_from_stringdbaccessory}
\alias{create_annotation_from_stringdbaccessory}
\title{create_annotation_from_stringdbaccessory()}
\usage{
create_annotation_from_stringdbaccessory(accessory_info)
}
\arguments{
\item{accessory_info}{#protein_info file from stringdb defined in get_networkhubdata_stringdb <- only use this function inside the other}
}
\value{
accessory_info_df dataframe that assignes each protein_id to ensemble or gene_symbol
}
\description{
create_annotation_from_stringdbaccessory()
}
\examples{
# create_annotation_from_stringdbaccessory(accessory_info)
#TODO: what can I do here as ppi_funcoup is not defined?

}
