% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_utils.R
\name{fetch_NetworkHub}
\alias{fetch_NetworkHub}
\title{Title}
\usage{
fetch_NetworkHub(rname, update = TRUE, nh_cachedir = "NetworkHub", ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
\examples{
fetch_NetworkHub(rname = "STRINGDB_Homo sapiens_v12.0")

}
