% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-1_db_stringdb.R
\name{get_accessoryinfo_stringdb}
\alias{get_accessoryinfo_stringdb}
\title{Title}
\usage{
get_accessoryinfo_stringdb(species, version, cache = TRUE)
}
\arguments{
\item{cache}{}
}
\description{
Title
}
\examples{
human_acc <- get_accessoryinfo_stringdb(species = "Homo sapiens",
                                        version = "12.0")
}
