% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_funcoup_functions.R
\name{get_annotation_funcoup}
\alias{get_annotation_funcoup}
\title{get_annotation_funcoup ()}
\usage{
get_annotation_funcoup(ppi_funcoup, species)
}
\arguments{
\item{ppi_funcoup}{variable defined by ppis_funcoup in get_networkdata_funcoup()}

\item{species}{from which species does the data come from c( "A.thaliana", "B.subtilis", "B.taurus", "C.elegans","C.familiaris", "C.intestinalis", "D.melanogatser", "D.rerio", "E.coli", "G.gallus", "H.sapiens", "M.jannaschii", "M.musculus", "O.sativa", "P.falciparum", "R.norvegicus", "S.cerevisiae", "S.pombe", "S.scrofa", "S.solfataricus")}
}
\value{
ppi_funcoup
}
\description{
get_annotation_funcoup ()
}
\examples{
\dontrun{
 db_funcoup_df  <- get_annotation_funcoup( ppi_funcoup,
                                           species = "H.sapiens")
}
}
