% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_genemania_functions.R
\name{get_annotation_genemania}
\alias{get_annotation_genemania}
\title{get_annotation_genemania ()}
\usage{
get_annotation_genemania(ppi_genemania, species)
}
\arguments{
\item{ppi_genemania}{variable defined by ppis_genemania in get_networkdata_genemania()}

\item{species}{from which species does the data come from}
}
\value{
ppi_genemania

anno_df
}
\description{
get_annotation_genemania ()
}
\examples{
\dontrun{
db_genemania_df <- get_networkdata_genemania(species = "Homo_sapiens",
                                             version = "current",
                                             get_annotation = FALSE,
                                             add_annotation = FALSE
                                             )

db_genemania_anno_df <- get_annotation_genemania( ppi_genemania = db_genemania_df,
                                                  species = "Homo_sapiens"
                                                )
}
}
