% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_hint_functions.R
\name{get_annotation_hint}
\alias{get_annotation_hint}
\title{get_annotation_hint ()}
\usage{
get_annotation_hint(ppi_hint, species)
}
\arguments{
\item{ppi_hint}{variable defined by ppis_hint in get_networkdata_hint()}

\item{species}{from which species does the data come from}
}
\value{
anno_df
}
\description{
get_annotation_hint ()
}
\examples{
\dontrun{
db_hint_df <- get_networkdata_hint(species = "HomoSapiens",
                                   version = "2024-06",
                                   type = "binary",
                                   get_annotation = FALSE,
                                   add_annotation = FALSE
                                   )

db_hint_anno_df <- get_annotation_hint(ppi_hint = db_hint_df,
                                       species = "HomoSapiens")
}

}
