% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_hippie_functions.R
\name{get_annotation_hippie}
\alias{get_annotation_hippie}
\title{get_annotation_hippie()}
\usage{
get_annotation_hippie(ppi_hippie, species, version)
}
\arguments{
\item{ppi_hippie}{variable defined by ppis_hippie in get_networkdata_hippie()}

\item{species}{default value = "Homo_sapiens", from which species does the data come from}

\item{version}{default value = "current", version of data files in HIPPIE}
}
\value{
anno_df (for corresponding species in hippie)
}
\description{
get_annotation_hippie()
}
\examples{

db_hippie_df <- get_networkdata_hippie(species = "Homo_sapiens",
                                       version = "current",
                                       cache = TRUE,
                                       get_annotation = FALSE,
                                       add_annotation = FALSE
                                       )

db_hippie_anno_df <- get_annotation_hippie(ppi_hippie = db_hippie_df,
                                           species = "Homo_sapiens",
                                           version = "current"
                                           )

# returns a dataframe containing #TODO rows and #TODO columns


}
