% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_huri_functions.R
\name{get_annotation_huri}
\alias{get_annotation_huri}
\title{get_annotation_huri ()}
\usage{
get_annotation_huri(ppi_huri, species)
}
\arguments{
\item{ppi_huri}{variable defined by ppis_huri in get_networkdata_huri()}

\item{species}{from which species does the data come from}
}
\value{
anno_df

ppi_huri
}
\description{
get_annotation_huri ()
}
\examples{
\dontrun{
db_huri_df <- get_networkdata_huri(
  species = "human",
  type = "HI-union",
  cache = TRUE,
  get_annotation = FALSE,
  add_annotation = FALSE
)

db_huri_anno_df <- get_annotation_huri(
  ppi_huri = db_huri_df,
  species = "human"
)
}
}
