% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_iid_functions.R
\name{get_annotation_iid}
\alias{get_annotation_iid}
\title{get_annotation_iid ()}
\usage{
get_annotation_iid(ppi_iid, species)
}
\arguments{
\item{ppi_iid}{variable defined by ppis_iid in get_networkdata_iid()}

\item{species}{from which species does the data come from}

\item{version}{version of the data files in iid}
}
\value{
ppi_iid

anno_df
}
\description{
get_annotation_iid ()
}
\examples{
\dontrun{
db_iid_df <- get_networkdata_iid(species = "human",
                                 version = "2021-05",
                                 get_annotation = FALSE,
                                 add_annotation = FALSE
                                )

db_iid_anno_df <- get_annotation_iid( ppi_iid = db_iid_df,
                                      species = "human"
                                      )
}
}
