% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_innatedb_functions.R
\name{get_annotation_innatedb}
\alias{get_annotation_innatedb}
\title{get_annotation_innatedb ()}
\usage{
get_annotation_innatedb(ppi_innatedb, species)
}
\arguments{
\item{ppi_innatedb}{variable defined by ppis_innatedb in get_networkdata_innatedb()}

\item{species}{from which species does the data come from}
}
\value{
ppi_innatedb

anno_df
}
\description{
get_annotation_innatedb ()
}
\examples{
\dontrun{
db_innatedb_df <- get_networkdata_innatedb(
  species = "taxid:9606(Human)",
  version = "5.4",
  cache = TRUE,
  get_annotation = FALSE,
  add_annotation = FALSE
)

db_innatedb_anno_df <- get_annotation_innatedb(
  ppi_innatedb = db_innatedb_df,
  species = "taxid:9606(Human)"
)
}
}
