% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_intact_functions.R
\name{get_annotation_intact}
\alias{get_annotation_intact}
\title{get_annotation_intact ()}
\usage{
get_annotation_intact(ppi_intact, species)
}
\arguments{
\item{ppi_intact}{variable defined by ppis_intact in get_networkdata_intact()}

\item{species}{from which species does the data come from}
}
\value{
ppi_intact
}
\description{
get_annotation_intact ()
}
\examples{
\dontrun{
db_intact_df <- get_networkdata_intact( species = "taxid:9606(human)|taxid:9606(Homo sapiens)",
                                        version = "current",
                                        cache = TRUE,
                                        get_annotation = FALSE,
                                        add_annotation = FALSE
                                        )

db_intact_anno_df <- get_annotation_intact( ppi_intact = db_intact_df,
                                            species = "taxid:9606(human)|taxid:9606(Homo sapiens)"
                                            )
}
}
