% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_irefindex_functions.R
\name{get_annotation_irefindex}
\alias{get_annotation_irefindex}
\title{get_annotation_irefindex ()}
\usage{
get_annotation_irefindex(ppi_irefindex, species)
}
\arguments{
\item{ppi_irefindex}{variable defined by ppis_irefindex in get_networkdata_irefindex()}

\item{species}{from which species does the data come from}
}
\value{
ppi_irefindex
}
\description{
get_annotation_irefindex ()
}
\examples{
\dontrun{
db_irefindex_df <- get_networkdata_irefindex(species = "Homo sapiens",
                                             version = "08-28-2023",
                                             get_annotation = FALSE,
                                             add_annotation = FALSE
                                             )

db_irefindex_anno_df <- get_annotation_irefindex( ppi_irefindex = db_irefindex_df,
                                                  species = "Homo sapiens"
                                                )
}
}
