% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_pathwaycommons_functions.R
\name{get_annotation_pathwaycommons}
\alias{get_annotation_pathwaycommons}
\title{get_annotation_pathwaycommons ()}
\usage{
get_annotation_pathwaycommons(ppi_pathwaycommons, species)
}
\arguments{
\item{ppi_pathwaycommons}{variable defined by ppis_pathwaycommons in get_networkdata_pathwaycommons()}

\item{species}{from which species does the data come from}
}
\value{
ppi_pathwaycommons
}
\description{
get_annotation_pathwaycommons ()
}
\examples{
\dontrun{
db_pathwaycommons_df <- get_networkdata_pathwaycommons( species = "human",
                                                        version = "v12",
                                                        cache = TRUE,
                                                        get_annotation = FALSE,
                                                        add_annotation = FALSE
                                                        )

db_pathwaycommons_anno_df <- get_annotation_pathwaycommons( ppi_pathwaycommons = db_pathwaycommons_df,
                                                            species = "human")
}
}
