% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_reactome_functions.R
\name{get_annotation_reactome}
\alias{get_annotation_reactome}
\title{get_annotation_reactome ()}
\usage{
get_annotation_reactome(ppi_reactome, species)
}
\arguments{
\item{ppi_reactome}{variable defined by ppis_reactome in get_networkdata_reactome()}

\item{species}{from which species does the data come from}
}
\value{
ppi_reactome
}
\description{
get_annotation_reactome ()
}
\examples{
\dontrun{
db_reactome_df <- get_networkdata_reactome( species = "taxid:9606(Homo sapiens)",
                                            version = "current",
                                            cache = TRUE,
                                            get_annotation = FALSE,
                                            add_annotation = FALSE
                                           )

db_reactome_anno_df <- get_annotation_reactome( ppi_reactome = db_reactome_df,
                                                species = "taxid:9606(Homo sapiens)"
                                                )
}
}
