% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_stringdb_functions.R
\name{get_annotation_stringdb}
\alias{get_annotation_stringdb}
\title{get_annotation_stringdb()}
\usage{
get_annotation_stringdb(species, version, cache = TRUE)
}
\arguments{
\item{species}{from which species does the data come from}

\item{version}{version of data files in STRING, default value = "12.0"}

\item{cache}{Default value set to TRUE (automatically checks if the data file is already stored in the cache)}
}
\value{
anno_df (for corresponding species in stringdb)
}
\description{
get_annotation_stringdb()
}
\examples{

db_stringdb_anno_df <- get_annotation_stringdb(
  species = "Homo sapiens",
  version = "12.0",
  cache = TRUE
)

# returns a dataframe containing 19699 rows and 5 columns

}
