% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-14_db_biogrid.R
\name{get_networkdata_biogrid}
\alias{get_networkdata_biogrid}
\title{get_networkdata_biogrid()}
\usage{
get_networkdata_biogrid(
  species = "9606",
  version = "4.4.238",
  cache = TRUE,
  ...
)
}
\arguments{
\item{species}{default value = "9606" - from which species does the data come from}

\item{version}{version of the data files in biogrid}

\item{cache}{default value set to TRUE (automatically checks if the data file is already stored in the cache)}

\item{...}{further arguments passed to or from other methods}
}
\value{
ppis_biogrid
}
\description{
get_networkdata_biogrid()
}
\examples{
\dontrun{
db_biogrid_df <- get_networkdata_biogrid(species = "9606",
                                           version = "4.4.238"
                                           )
db_biogrid_df
}
}
