% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-8_db_cpdb.R
\name{get_networkdata_cpdb}
\alias{get_networkdata_cpdb}
\title{get_networkdata_cpdb()}
\usage{
get_networkdata_cpdb(species = "human", cache = TRUE, ...)
}
\arguments{
\item{species}{from which species does the data come from (default human because currently only human data provided from cpdb)}

\item{cache}{default value set to TRUE (automatically checks if the data file is already stored in the cache)}

\item{...}{further arguments passed to or from other methods}
}
\value{
ppis_cpdb
}
\description{
get_networkdata_cpdb()
}
\examples{
\dontrun{
db_cpdb_df <- get_networkdata_cpdb(
  species = "human"
)

db_cpdb_df
}
}
