% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_funcoup_functions.R
\name{get_networkdata_funcoup}
\alias{get_networkdata_funcoup}
\title{get_networkdata_funcoup()}
\usage{
get_networkdata_funcoup(
  species = "H.sapiens",
  version = "6.0",
  type = c("compact", "full"),
  cache = TRUE,
  get_annotation = TRUE,
  add_annotation = TRUE,
  ...
)
}
\arguments{
\item{species}{from which species does the data come from c( "A.thaliana", "B.subtilis", "B.taurus", "C.elegans","C.familiaris", "C.intestinalis", "D.melanogatser", "D.rerio", "E.coli", "G.gallus", "H.sapiens", "M.jannaschii", "M.musculus", "O.sativa", "P.falciparum", "R.norvegicus", "S.cerevisiae", "S.pombe", "S.scrofa", "S.solfataricus")}

\item{version}{version of the data files in funcoup}

\item{type}{compact or full file}

\item{cache}{default value set to TRUE (automatically checks if the data file is already stored in the cache)}

\item{get_annotation}{creation of an annotation dataframe using AnnotationDbi packages, default value set to TRUE}

\item{add_annotation}{adding annotation to ppi dataframe, default value set to TRUE}

\item{...}{further arguments passed to or from other methods}
}
\value{
ppis_funcoup

ppis_annotated_funcoup
}
\description{
get_networkdata_funcoup()
}
\examples{
\dontrun{
db_funcoup_df <- get_networkdata_funcoup(species = "H.sapiens",
                                         version = "6.0",
                                         type = "compact",
                                         cache = TRUE,
                                         get_annotation = FALSE,
                                         add_annotation = FALSE
                                         )
db_funcoup_df
}

}
