% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_hint_functions.R
\name{get_networkdata_hint}
\alias{get_networkdata_hint}
\title{get_networkdata_hint()}
\usage{
get_networkdata_hint(
  species,
  version,
  type = "binary",
  cache = TRUE,
  get_annotation = TRUE,
  add_annotation = TRUE,
  ...
)
}
\arguments{
\item{species}{from which species does the data come from}

\item{version}{version of the data files in hint}

\item{type}{different interaction files provided by hint (all high-quality)}

\item{cache}{default value set to TRUE (automatically checks if the data file is already stored in the cache)}

\item{get_annotation}{creation of an annotation dataframe using AnnotationDbi packages, default value set to TRUE}

\item{add_annotation}{adding annotation to ppi dataframe, default value set to TRUE}

\item{...}{further arguments passed to or from other methods}
}
\value{
ppis_hint

db_hint_ppi_anno_df

db_hint_anno_df
}
\description{
get_networkdata_hint()
}
\examples{
\dontrun{
db_hint_df <- get_networkdata_hint(
  species = "HomoSapiens",
  version = "2024-06",
  type = "binary",
  cache = TRUE,
  get_annotation = FALSE,
  add_annotation = FALSE
)
}
}
