% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-9_db_huri.R
\name{get_networkdata_huri}
\alias{get_networkdata_huri}
\title{get_networkdata_huri()}
\usage{
get_networkdata_huri(
  species = "human",
  type = "HI-union",
  cache = TRUE,
  add_annotation = TRUE,
  ...
)
}
\arguments{
\item{species}{from which species does the data come from (default human because currently only human data provided from huri)}

\item{type}{different datasets , more information on "http://www.interactome-atlas.org/about/" , c("HI-union", "Lit-BM")}

\item{cache}{default value set to TRUE (automatically checks if the data file is already stored in the cache)}

\item{add_annotation}{expanding the dataframe with six columns (GeneSymbol, Uniprot ID and Entrez_ID)}

\item{...}{further arguments passed to or from other methods}
}
\value{
ppis_huri
}
\description{
get_networkdata_huri()
}
\examples{

db_huri_df <- get_networkdata_huri(
  species = "human",
  type = "HI-union"
)

db_huri_df

}
