% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_iid_functions.R
\name{get_networkdata_iid}
\alias{get_networkdata_iid}
\title{get_networkdata_iid()}
\usage{
get_networkdata_iid(
  species = "human",
  version = "2021-05",
  cache = TRUE,
  get_annotation = TRUE,
  add_annotation = TRUE,
  ...
)
}
\arguments{
\item{species}{from which species does the data come from}

\item{version}{version of the data files in iid}

\item{cache}{default value set to TRUE (automatically checks if the data file is already stored in the cache)}

\item{get_annotation}{creation of an annotation dataframe using AnnotationDbi packages, default value set to TRUE}

\item{add_annotation}{adding annotation to ppi dataframe, default value set to TRUE}

\item{...}{further arguments passed to or from other methods}
}
\value{
ppis_iid

db_iid_ppi_anno_df

db_iid_anno_df
}
\description{
get_networkdata_iid()
}
\examples{
\dontrun{
db_iid_df <- get_networkdata_iid(
  species = "human",
  version = "2021-05",
  get_annotation = FALSE,
  add_annotation = FALSE
)
}
}
