% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-15_db_intact.R
\name{get_networkdata_intact}
\alias{get_networkdata_intact}
\title{get_networkdata_intact()}
\usage{
get_networkdata_intact(
  species = "taxid:9606(human)|taxid:9606(Homo sapiens)",
  version = "current",
  cache = TRUE,
  add_annotation = TRUE,
  ...
)
}
\arguments{
\item{species}{default value = "taxid:9606(human)|taxid:9606(Homo sapiens)" - from which species does the data come from}

\item{version}{default value = "current" version of the data files in intact}

\item{cache}{default value set to TRUE (automatically checks if the data file is already stored in the cache)}

\item{add_annotation}{expanding the dataframe with six columns (GeneSymbol,Entrez_ID and Ensembl_ID)}

\item{...}{further arguments passed to or from other methods}
}
\value{
ppis_intact
}
\description{
get_networkdata_intact()
}
\examples{
\donttest{
db_intact_df <- get_networkdata_intact(species = "taxid:9606(human)|taxid:9606(Homo sapiens)",
                                           version =  "current"
                                           )
db_intact_df
}
}
