% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-5_db_irefindex.R
\name{get_networkdata_irefindex}
\alias{get_networkdata_irefindex}
\title{get_networkdata_irefindex()}
\usage{
get_networkdata_irefindex(
  species,
  version = "08-28-2023",
  cache = TRUE,
  add_annotation = TRUE,
  ...
)
}
\arguments{
\item{species}{from which species does the data come from}

\item{version}{version of the data files in irefindex}

\item{cache}{default value set to TRUE (automatically checks if the data file is already stored in the cache)}

\item{...}{further arguments passed to or from other methods}
}
\value{
ppis_irefindex
}
\description{
get_networkdata_irefindex()
}
\examples{

db_irefindex_df <- get_networkdata_irefindex(
  species = "Homo sapiens",
  version = "08-28-2023")

db_irefindex_df

}
