% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-10_db_matrixdb.R
\name{get_networkdata_matrixdb}
\alias{get_networkdata_matrixdb}
\title{get_networkdata_matrixdb()}
\usage{
get_networkdata_matrixdb(
  species = "human",
  type = "CORE",
  cache = TRUE,
  add_annotation = TRUE,
  ...
)
}
\arguments{
\item{species}{from which species does the data come from (default human because currently only human data provided from matrixdb)}

\item{type}{datasets provided by MatrixDB: "CORE" = MatrixDB manually curated interaction dataset}

\item{cache}{default value set to TRUE (automatically checks if the data file is already stored in the cache)}

\item{add_annotation}{expanding the dataframe with (GeneSymbol, Uniprot ID and Entrez_ID)}

\item{...}{further arguments passed to or from other methods}
}
\value{
ppis_matrixdb
}
\description{
get_networkdata_matrixdb()
}
\examples{

db_matrixdb_df <- get_networkdata_matrixdb(
  species = "human",
  type = "CORE"
)

db_matrixdb_df

}
