% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_mint_functions.R
\name{get_networkdata_mint}
\alias{get_networkdata_mint}
\title{get_networkdata_mint()}
\usage{
get_networkdata_mint(
  species,
  version = "current",
  cache = TRUE,
  add_annotation = TRUE,
  ...
)
}
\arguments{
\item{species}{from which species does the data come from}

\item{version}{version of the data files in mint}

\item{cache}{default value set to TRUE (automatically checks if the data file is already stored in the cache)}

\item{add_annotation}{default value set to TRUE (MINT annotation dataframe contains Uniprot, GeneSymbol and Ensembl)}

\item{...}{further arguments passed to or from other methods}
}
\value{
ppis_mint
}
\description{
get_networkdata_mint()
}
\examples{
\dontrun{
db_mint_df <- get_networkdata_mint(
  species = "Homo Sapiens",
  version = "current"
)

db_mint_df
}
}
