% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-11_db_pathwaycommons.R
\name{get_networkdata_pathwaycommons}
\alias{get_networkdata_pathwaycommons}
\title{get_networkdata_pathwaycommons()}
\usage{
get_networkdata_pathwaycommons(
  species = "human",
  version = "v12",
  cache = TRUE,
  add_annotation = TRUE,
  ...
)
}
\arguments{
\item{species}{from which species does the data come from}

\item{version}{version of the data files in pathwaycommons}

\item{cache}{default value set to TRUE (automatically checks if the data file is already stored in the cache)}

\item{add_annotation}{expanding the dataframe with four columns (Entrez_ID and Ensembl_ID)}

\item{...}{further arguments passed to or from other methods}
}
\value{
ppis_pathwaycommons
}
\description{
get_networkdata_pathwaycommons()
}
\examples{
\dontrun{
db_pathwaycommons_df <- get_networkdata_pathwaycommons(
  species = "human",
  version = "v12"
)

db_pathwaycommons_df
}
}
