% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-12_db_reactome.R
\name{get_networkdata_reactome}
\alias{get_networkdata_reactome}
\title{get_networkdata_reactome()}
\usage{
get_networkdata_reactome(
  species = "taxid:9606(Homo sapiens)",
  version = "current",
  cache = TRUE,
  add_annotation = TRUE,
  ...
)
}
\arguments{
\item{species}{default value = "taxid:9606(Homo sapiens)" - from which species does the data come from}

\item{version}{version of the data files in reactome}

\item{cache}{default value set to TRUE (automatically checks if the data file is already stored in the cache)}

\item{add_annotation}{expanding the dataframe with six columns (GeneSymbol,Entrez_ID and Ensembl_ID)}

\item{...}{further arguments passed to or from other methods}
}
\value{
ppis_reactome
}
\description{
get_networkdata_reactome()
}
\examples{
\dontrun{
db_reactome_df <- get_networkdata_reactome(species = "taxid:9606(Homo sapiens)",
                                           version = "current"
                                           )
db_reactome_df
}
}
