% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-1_db_stringdb.R
\name{get_networkdata_stringdb}
\alias{get_networkdata_stringdb}
\title{Title}
\usage{
get_networkdata_stringdb(
  species,
  version,
  cache = TRUE,
  remap_identifiers = TRUE,
  remap_to = c("ENSEMBL", "gene_symbol"),
  ...
)
}
\arguments{
\item{species}{from which species does the data come from}

\item{version}{version of the data files in stringdb}

\item{cache}{efault value set to TRUE (automatically checks if the data file is already stored in the cache)}

\item{remap_identifiers}{TODO}

\item{remap_to}{TODO}

\item{...}{TODO}
}
\value{
ppis_stringdb TODO
}
\description{
Title
}
\examples{

db_string_df <- get_networkdata_stringdb(
  species = "Homo sapiens",
  version = "12.0"
)

db_string_df

}
