% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-1_db_stringdb.R
\name{info_species_stringdb}
\alias{info_species_stringdb}
\title{info_species_stringdb}
\usage{
info_species_stringdb(version = "12.0")
}
\arguments{
\item{version}{version of the data files in stringdb}
}
\value{
df_species returns the file containing info about the species annotation
}
\description{
info_species_stringdb
}
\examples{
info_species_stringdb()

}
