% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_utils.R
\name{initialize_NetworkHub}
\alias{initialize_NetworkHub}
\title{initialize_NetworkHub function}
\usage{
initialize_NetworkHub(nh_cachedir = "NetworkHub")
}
\arguments{
\item{nh_cachedir}{}
}
\description{
initialize_NetworkHub function
}
\examples{
library("BiocFileCache") # load the library of the package
bfc_nh <- initialize_NetworkHub() # use this function to create the cache called "bfc_nh"
bfccache(bfc_nh) # get the path to the cache
length(bfc_nh) # how many entries?
bfcinfo(bfc_nh) # what are the column names of the cache
}
