% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_utils.R
\name{map_SE_on_graph}
\alias{map_SE_on_graph}
\title{map_SE_on_graph}
\usage{
map_SE_on_graph(se, g)
}
\arguments{
\item{se}{summerized experiment object (e.g. DE analysis resluts)}

\item{g}{igraph object (containing ppi interactions from corresponding database)}
}
\description{
map_SE_on_graph
}
\examples{

data(res_de_macrophage, package = "GeneTonic")
res_de <- res_macrophage_IFNg_vs_naive

db_stringdb_df <- get_networkdata_stringdb(species = "Homo sapiens",
                                           version = "12.0"
                                           )

db_stringdb_anno_df  <- annotation_stringdb(ppi_stringdb = db_stringdb_df,
                                            species = "Homo sapiens",
                                            version = "12.0",
                                            create_ppi_anno_df = FALSE)

db_stringdb_igraph_object <- build_graph(graph_data = db_stringdb_df,
                                         graph_data_anno = db_stringdb_anno_df,
                                         data_source = "stringdb",
                                         output_format = "igraph",
                                         min_score_threshold = "0.35",
                                         add_info_nodes = TRUE)


map_SE_on_graph(se = res_de,
                g = db_stringdb_igraph_object
                )

}
