% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkHub-pkg.R
\docType{package}
\name{NetworkHub-pkg}
\alias{NetworkHub}
\alias{NetworkHub-package}
\alias{NetworkHub-pkg}
\title{NetworkHub}
\description{
NetworkHub
}
\details{
This package provides functionality to ... TODO
}
\author{
\strong{Maintainer}: Lotta Wagner \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

Authors:
\itemize{
  \item First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})
}

}
\keyword{internal}
