% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urlmaker_functions.R
\name{urlmaker_biogrid}
\alias{urlmaker_biogrid}
\title{URL maker for BioGRID}
\usage{
urlmaker_biogrid(version = "4.4.238")
}
\arguments{
\item{version}{version of the data files in BioGRID}
}
\value{
A character string with the URL as requested
}
\description{
Creating the URL to access the resources on \href{https://thebiogrid.org/}{BioGRID}
}
\examples{
url_biogrid <- urlmaker_biogrid()

url_biogrid
}
\references{
https://thebiogrid.org/
}
\seealso{
Other urlmakers: 
\code{\link{urlmaker_cpdb}()},
\code{\link{urlmaker_funcoup}()},
\code{\link{urlmaker_genemania}()},
\code{\link{urlmaker_hint}()},
\code{\link{urlmaker_hippie}()},
\code{\link{urlmaker_huri}()},
\code{\link{urlmaker_iid}()},
\code{\link{urlmaker_innatedb}()},
\code{\link{urlmaker_intact}()},
\code{\link{urlmaker_irefindex}()},
\code{\link{urlmaker_matrixdb}()},
\code{\link{urlmaker_mint}()},
\code{\link{urlmaker_pathwaycommons}()},
\code{\link{urlmaker_reactome}()},
\code{\link{urlmaker_stringdb}()}
}
\concept{urlmakers}
