% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_urlmaker.R
\name{urlmaker_corum}
\alias{urlmaker_corum}
\title{urlmaker_corum()}
\usage{
urlmaker_corum(version = "current")
}
\arguments{
\item{version}{version of the data files in CORUM, default value = "current" (August 2024 = 28.11.2022 Corum 4.1 release)}
}
\value{
url returns the corresponding url set by params #SPECIESDEFINITION later on
}
\description{
urlmaker_corum()
}
\examples{
urlmaker_corum()
}
