% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urlmaker_functions.R
\name{urlmaker_cpdb}
\alias{urlmaker_cpdb}
\title{URL maker for ConsensusPathDB}
\usage{
urlmaker_cpdb(species = "human")
}
\arguments{
\item{species}{Character string, specifying the species of interesting,
defaulting to  "human".}
}
\value{
A character string with the URL as requested
}
\description{
Creating the URL to access the resources on
\href{http://cpdb.molgen.mpg.de/}{ConsensusPathDB}
}
\examples{
url_cpdb <- urlmaker_cpdb(species = "human")

url_cpdb
}
\references{
http://cpdb.molgen.mpg.de/
}
\seealso{
Other urlmakers: 
\code{\link{urlmaker_biogrid}()},
\code{\link{urlmaker_funcoup}()},
\code{\link{urlmaker_genemania}()},
\code{\link{urlmaker_hint}()},
\code{\link{urlmaker_hippie}()},
\code{\link{urlmaker_huri}()},
\code{\link{urlmaker_iid}()},
\code{\link{urlmaker_innatedb}()},
\code{\link{urlmaker_intact}()},
\code{\link{urlmaker_irefindex}()},
\code{\link{urlmaker_matrixdb}()},
\code{\link{urlmaker_mint}()},
\code{\link{urlmaker_pathwaycommons}()},
\code{\link{urlmaker_reactome}()},
\code{\link{urlmaker_stringdb}()}
}
\concept{urlmakers}
