% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urlmaker_functions.R
\name{urlmaker_funcoup}
\alias{urlmaker_funcoup}
\title{URL maker for FunCoup}
\usage{
urlmaker_funcoup(
  version = "6.0",
  species = "H.sapiens",
  type = c("compact", "full")
)
}
\arguments{
\item{version}{Character string, specifying the version of FunCoup to
retrieve, default value = "6.0"}

\item{species}{Character string, types listed in \code{list_species_funcoup}
depending on current version. Defaults to "H.sapiens"}

\item{type}{Character string, specifying the interaction types in FunCoup
(for current version compact recommended, for older versions only full possible)}
}
\value{
A character string with the URL as requested
}
\description{
Creating the URL to access the resources on \href{https://funcoup.org/}{FunCoup}
}
\examples{
url_funcoup <- urlmaker_funcoup(
  version = "6.0",
  species = "B.taurus",
  type = "full"
)

url_funcoup
}
\references{
https://funcoup.org/
}
\seealso{
Other urlmakers: 
\code{\link{urlmaker_biogrid}()},
\code{\link{urlmaker_cpdb}()},
\code{\link{urlmaker_genemania}()},
\code{\link{urlmaker_hint}()},
\code{\link{urlmaker_hippie}()},
\code{\link{urlmaker_huri}()},
\code{\link{urlmaker_iid}()},
\code{\link{urlmaker_innatedb}()},
\code{\link{urlmaker_intact}()},
\code{\link{urlmaker_irefindex}()},
\code{\link{urlmaker_matrixdb}()},
\code{\link{urlmaker_mint}()},
\code{\link{urlmaker_pathwaycommons}()},
\code{\link{urlmaker_reactome}()},
\code{\link{urlmaker_stringdb}()}
}
\concept{urlmakers}
