% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urlmaker_functions.R
\name{urlmaker_genemania}
\alias{urlmaker_genemania}
\title{URL maker for GeneMania}
\usage{
urlmaker_genemania(species = "Homo_sapiens", version = "current")
}
\arguments{
\item{species}{Character string, types listed in \code{list_species_genemania}
depending on current version, default value = "Homo_sapiens"}

\item{version}{Character string, version of GeneMania , defaults to "current"}
}
\value{
A character string with the URL as requested
}
\description{
Creating the URL to access the resources on \href{https://genemania.org/}{GeneMania}
}
\examples{
url_genemania <- urlmaker_genemania(
  species = "Homo_sapiens",
  version = "current"
)

url_genemania
}
\references{
https://genemania.org/
}
\seealso{
Other urlmakers: 
\code{\link{urlmaker_biogrid}()},
\code{\link{urlmaker_cpdb}()},
\code{\link{urlmaker_funcoup}()},
\code{\link{urlmaker_hint}()},
\code{\link{urlmaker_hippie}()},
\code{\link{urlmaker_huri}()},
\code{\link{urlmaker_iid}()},
\code{\link{urlmaker_innatedb}()},
\code{\link{urlmaker_intact}()},
\code{\link{urlmaker_irefindex}()},
\code{\link{urlmaker_matrixdb}()},
\code{\link{urlmaker_mint}()},
\code{\link{urlmaker_pathwaycommons}()},
\code{\link{urlmaker_reactome}()},
\code{\link{urlmaker_stringdb}()}
}
\concept{urlmakers}
