% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urlmaker_functions.R
\name{urlmaker_huri}
\alias{urlmaker_huri}
\title{URL maker for HuRI}
\usage{
urlmaker_huri(species = "human", type = c("HI-union", "Lit-BM"))
}
\arguments{
\item{species}{Character string, default value = "human", because this
database only provides human data}

\item{type}{Character string, different datasets , more information on
"http://www.interactome-atlas.org/about/" - defaults to "HI-union"}
}
\value{
A character string with the URL as requested
}
\description{
Creating the URL to access the resources on
\href{http://www.interactome-atlas.org/}{HuRI} (The Human Reference Protein
Interactome)
}
\examples{

url_huri <- urlmaker_huri(
  species = "human",
  type = "HI-union"
)
url_huri
}
\references{
http://www.interactome-atlas.org/
}
\seealso{
Other urlmakers: 
\code{\link{urlmaker_biogrid}()},
\code{\link{urlmaker_cpdb}()},
\code{\link{urlmaker_funcoup}()},
\code{\link{urlmaker_genemania}()},
\code{\link{urlmaker_hint}()},
\code{\link{urlmaker_hippie}()},
\code{\link{urlmaker_iid}()},
\code{\link{urlmaker_innatedb}()},
\code{\link{urlmaker_intact}()},
\code{\link{urlmaker_irefindex}()},
\code{\link{urlmaker_matrixdb}()},
\code{\link{urlmaker_mint}()},
\code{\link{urlmaker_pathwaycommons}()},
\code{\link{urlmaker_reactome}()},
\code{\link{urlmaker_stringdb}()}
}
\concept{urlmakers}
