% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urlmaker_functions.R
\name{urlmaker_iid}
\alias{urlmaker_iid}
\title{URL maker for IID}
\usage{
urlmaker_iid(species = "human", version = "2021-05")
}
\arguments{
\item{species}{Character string, with species listed in \code{list_species_iid}
depending on current version, default value = "human"}

\item{version}{Character string, version of IID , default value = "2021-05"}
}
\value{
A character string with the URL as requested
}
\description{
Creating the URL to access the resources on \href{https://iid.ophid.utoronto.ca/}{IID}
(Integrated Interactions Database)
}
\examples{
url_iid <- urlmaker_iid(species = "mouse")

url_iid
}
\references{
https://iid.ophid.utoronto.ca/
}
\seealso{
Other urlmakers: 
\code{\link{urlmaker_biogrid}()},
\code{\link{urlmaker_cpdb}()},
\code{\link{urlmaker_funcoup}()},
\code{\link{urlmaker_genemania}()},
\code{\link{urlmaker_hint}()},
\code{\link{urlmaker_hippie}()},
\code{\link{urlmaker_huri}()},
\code{\link{urlmaker_innatedb}()},
\code{\link{urlmaker_intact}()},
\code{\link{urlmaker_irefindex}()},
\code{\link{urlmaker_matrixdb}()},
\code{\link{urlmaker_mint}()},
\code{\link{urlmaker_pathwaycommons}()},
\code{\link{urlmaker_reactome}()},
\code{\link{urlmaker_stringdb}()}
}
\concept{urlmakers}
