% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urlmaker_functions.R
\name{urlmaker_innatedb}
\alias{urlmaker_innatedb}
\title{URL maker for InnateDB}
\usage{
urlmaker_innatedb(url = url, version = "5.4")
}
\arguments{
\item{url}{Character string, innatedb doesn't provide information about the version or the species in their url, that why the default value of url is url}

\item{version}{Character string, parameter set to current version 5.4 #UPDATEVERSION}
}
\value{
A character string with the URL as requested
}
\description{
Creating the URL to access the resources on
\href{https://www.innatedb.com/}{InnateDB}
}
\examples{
url_innatedb <- urlmaker_innatedb()

url_innatedb
}
\references{
https://www.innatedb.com/
}
\seealso{
Other urlmakers: 
\code{\link{urlmaker_biogrid}()},
\code{\link{urlmaker_cpdb}()},
\code{\link{urlmaker_funcoup}()},
\code{\link{urlmaker_genemania}()},
\code{\link{urlmaker_hint}()},
\code{\link{urlmaker_hippie}()},
\code{\link{urlmaker_huri}()},
\code{\link{urlmaker_iid}()},
\code{\link{urlmaker_intact}()},
\code{\link{urlmaker_irefindex}()},
\code{\link{urlmaker_matrixdb}()},
\code{\link{urlmaker_mint}()},
\code{\link{urlmaker_pathwaycommons}()},
\code{\link{urlmaker_reactome}()},
\code{\link{urlmaker_stringdb}()}
}
\concept{urlmakers}
