% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urlmaker_functions.R
\name{urlmaker_intact}
\alias{urlmaker_intact}
\title{URL maker for IntAct}
\usage{
urlmaker_intact(version = "current")
}
\arguments{
\item{version}{Character string, version of the data files in IntAct,
default value = "current" (August 2024  = 2024-05-23 18:09    6.6G)}
}
\value{
A character string with the URL as requested
}
\description{
Creating the URL to access the resources on
\href{https://www.ebi.ac.uk/intact/}{IntAct}
}
\examples{
url_intact <- urlmaker_intact(version = "current")

url_intact
}
\references{
https://www.ebi.ac.uk/intact/
}
\seealso{
Other urlmakers: 
\code{\link{urlmaker_biogrid}()},
\code{\link{urlmaker_cpdb}()},
\code{\link{urlmaker_funcoup}()},
\code{\link{urlmaker_genemania}()},
\code{\link{urlmaker_hint}()},
\code{\link{urlmaker_hippie}()},
\code{\link{urlmaker_huri}()},
\code{\link{urlmaker_iid}()},
\code{\link{urlmaker_innatedb}()},
\code{\link{urlmaker_irefindex}()},
\code{\link{urlmaker_matrixdb}()},
\code{\link{urlmaker_mint}()},
\code{\link{urlmaker_pathwaycommons}()},
\code{\link{urlmaker_reactome}()},
\code{\link{urlmaker_stringdb}()}
}
\concept{urlmakers}
