% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urlmaker_functions.R
\name{urlmaker_irefindex}
\alias{urlmaker_irefindex}
\title{URL maker for iRefIndex}
\usage{
urlmaker_irefindex(species, version = "08-28-2023")
}
\arguments{
\item{species}{Character string, from which species does the data come from,
default value = "Homo sapiens"}

\item{version}{Character string, version of data files in iRefIndex,
default value = "08-28-2023" #UPDATEVERSION}
}
\value{
A character string with the URL as requested
}
\description{
Creating the URL to access the resources on \href{https://irefindex.vib.be/}{iRefIndex}
(Integrated Interactions Database)
}
\examples{
url_irefindex <- urlmaker_irefindex(
  species = "Homo sapiens",
  version = "08-28-2023"
)
url_irefindex
}
\references{
https://irefindex.vib.be/
}
\seealso{
Other urlmakers: 
\code{\link{urlmaker_biogrid}()},
\code{\link{urlmaker_cpdb}()},
\code{\link{urlmaker_funcoup}()},
\code{\link{urlmaker_genemania}()},
\code{\link{urlmaker_hint}()},
\code{\link{urlmaker_hippie}()},
\code{\link{urlmaker_huri}()},
\code{\link{urlmaker_iid}()},
\code{\link{urlmaker_innatedb}()},
\code{\link{urlmaker_intact}()},
\code{\link{urlmaker_matrixdb}()},
\code{\link{urlmaker_mint}()},
\code{\link{urlmaker_pathwaycommons}()},
\code{\link{urlmaker_reactome}()},
\code{\link{urlmaker_stringdb}()}
}
\concept{urlmakers}
