% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urlmaker_functions.R
\name{urlmaker_matrixdb}
\alias{urlmaker_matrixdb}
\title{URL maker for MatrixDB}
\usage{
urlmaker_matrixdb(species = "human", type = c("all", "CORE"), version = "4_0")
}
\arguments{
\item{species}{Character string, default value = "human", because only one
version and one species at MatrixDB #UPDATEVERSION}

\item{type}{Character string, datasets provided by MatrixDB: "CORE"
= MatrixDB manually curated interaction dataset. Defaults to "all"}

\item{version}{Character string, specifying the version number. Defaults to
the recently updated "4.0" (as this was recently added to the URL)}
}
\value{
A character string with the URL as requested
}
\description{
Creating the URL to access the resources on
\href{https://matrixdb.univ-lyon1.fr/}{MatrixDB}
}
\examples{
url_matrixdb <- urlmaker_matrixdb(type = "CORE")

url_matrixdb
}
\references{
https://matrixdb.univ-lyon1.fr/
}
\seealso{
Other urlmakers: 
\code{\link{urlmaker_biogrid}()},
\code{\link{urlmaker_cpdb}()},
\code{\link{urlmaker_funcoup}()},
\code{\link{urlmaker_genemania}()},
\code{\link{urlmaker_hint}()},
\code{\link{urlmaker_hippie}()},
\code{\link{urlmaker_huri}()},
\code{\link{urlmaker_iid}()},
\code{\link{urlmaker_innatedb}()},
\code{\link{urlmaker_intact}()},
\code{\link{urlmaker_irefindex}()},
\code{\link{urlmaker_mint}()},
\code{\link{urlmaker_pathwaycommons}()},
\code{\link{urlmaker_reactome}()},
\code{\link{urlmaker_stringdb}()}
}
\concept{urlmakers}
