% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urlmaker_functions.R
\name{urlmaker_mint}
\alias{urlmaker_mint}
\title{URL maker for MINT}
\usage{
urlmaker_mint(species = "Homo Sapiens", version = "current")
}
\arguments{
\item{species}{Character string, from which species does the data come from,
default value = "Homo Sapiens"}

\item{version}{Character string, version of data files in MINT, defaults to
"current" (as it is specified by the MINT curators)}
}
\value{
A character string with the URL as requested
}
\description{
Creating the URL to access the resources on \href{https://mint.bio.uniroma2.it/}{MINT}
(The Molecular INTeraction Database)
}
\examples{
url_mint <- urlmaker_mint(species = "Homo Sapiens")
url_mint
}
\references{
https://mint.bio.uniroma2.it/
}
\seealso{
Other urlmakers: 
\code{\link{urlmaker_biogrid}()},
\code{\link{urlmaker_cpdb}()},
\code{\link{urlmaker_funcoup}()},
\code{\link{urlmaker_genemania}()},
\code{\link{urlmaker_hint}()},
\code{\link{urlmaker_hippie}()},
\code{\link{urlmaker_huri}()},
\code{\link{urlmaker_iid}()},
\code{\link{urlmaker_innatedb}()},
\code{\link{urlmaker_intact}()},
\code{\link{urlmaker_irefindex}()},
\code{\link{urlmaker_matrixdb}()},
\code{\link{urlmaker_pathwaycommons}()},
\code{\link{urlmaker_reactome}()},
\code{\link{urlmaker_stringdb}()}
}
\concept{urlmakers}
