% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urlmaker_functions.R
\name{urlmaker_pathwaycommons}
\alias{urlmaker_pathwaycommons}
\title{URL maker for PathwayCommons}
\usage{
urlmaker_pathwaycommons(species = "human", version = "v12")
}
\arguments{
\item{species}{Character string, default value = "Homo sapiens", because
only one species is available}

\item{version}{Character string, default value = "v12" (v14 doesn't contain
all datafiles as of Feb 2025)}
}
\value{
A character string with the URL as requested
}
\description{
Creating the URL to access the resources on
\href{https://www.pathwaycommons.org/}{PathwayCommons}
}
\examples{

url_pc <- urlmaker_pathwaycommons()

url_pc
}
\references{
https://www.pathwaycommons.org/
}
\seealso{
Other urlmakers: 
\code{\link{urlmaker_biogrid}()},
\code{\link{urlmaker_cpdb}()},
\code{\link{urlmaker_funcoup}()},
\code{\link{urlmaker_genemania}()},
\code{\link{urlmaker_hint}()},
\code{\link{urlmaker_hippie}()},
\code{\link{urlmaker_huri}()},
\code{\link{urlmaker_iid}()},
\code{\link{urlmaker_innatedb}()},
\code{\link{urlmaker_intact}()},
\code{\link{urlmaker_irefindex}()},
\code{\link{urlmaker_matrixdb}()},
\code{\link{urlmaker_mint}()},
\code{\link{urlmaker_reactome}()},
\code{\link{urlmaker_stringdb}()}
}
\concept{urlmakers}
