% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_urlmaker.R
\name{urlmaker_pc}
\alias{urlmaker_pc}
\title{urlmaker_pc()}
\usage{
urlmaker_pc(species = "Homo sapiens", version = "v12")
}
\arguments{
\item{species}{default value = "Homo sapiens", because only one species}

\item{version}{default value = "v12" because v14 doesn't contain all datafiles #UPDATEVERSION}
}
\value{
url returns the corresponding url set by params
}
\description{
urlmaker_pc()
}
\examples{

urlmaker_pc()

}
