% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urlmaker_functions.R
\name{urlmaker_reactome}
\alias{urlmaker_reactome}
\title{URL maker for Reactome}
\usage{
urlmaker_reactome(version = "current")
}
\arguments{
\item{version}{Character string, default value = "current" #UPDATEVERSION}
}
\value{
A character string with the URL as requested
}
\description{
Creating the URL to access the resources on \href{https://reactome.org/}{Reactome}
}
\examples{
url_reactome <- urlmaker_reactome()

url_reactome
}
\references{
https://reactome.org/
}
\seealso{
Other urlmakers: 
\code{\link{urlmaker_biogrid}()},
\code{\link{urlmaker_cpdb}()},
\code{\link{urlmaker_funcoup}()},
\code{\link{urlmaker_genemania}()},
\code{\link{urlmaker_hint}()},
\code{\link{urlmaker_hippie}()},
\code{\link{urlmaker_huri}()},
\code{\link{urlmaker_iid}()},
\code{\link{urlmaker_innatedb}()},
\code{\link{urlmaker_intact}()},
\code{\link{urlmaker_irefindex}()},
\code{\link{urlmaker_matrixdb}()},
\code{\link{urlmaker_mint}()},
\code{\link{urlmaker_pathwaycommons}()},
\code{\link{urlmaker_stringdb}()}
}
\concept{urlmakers}
