% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urlmaker_functions.R
\name{urlmaker_stringdb}
\alias{urlmaker_stringdb}
\title{URL maker for STRINGDB}
\usage{
urlmaker_stringdb(
  type = c("PPI", "protein_info"),
  species = "Homo sapiens",
  version = "12.0"
)
}
\arguments{
\item{type}{Character string, which data file do you want to download?
Can be one of "PPI" or "protein_info"}

\item{species}{Character string, from which species does the data come from.
Defaults to "Homo sapiens" for human}

\item{version}{Character string, specifying the version of the data files in
stringdb - defaults to "12.0" for the latest release (as of Feb 2025)}
}
\value{
A character string with the URL as requested
}
\description{
Creating the URL to access the resources on \href{https://string-db.org/}{STRINGDB}
}
\details{
The urlmaker function will automatically handle the conversion between human
readable identifiers and the identifier used internally by the selected resource
}
\examples{
url_stringdb <- urlmaker_stringdb(
  type = "PPI",
  species = "Homo sapiens",
  version = "12.0"
)

url_stringdb
}
\references{
https://string-db.org/
}
\seealso{
Other urlmakers: 
\code{\link{urlmaker_biogrid}()},
\code{\link{urlmaker_cpdb}()},
\code{\link{urlmaker_funcoup}()},
\code{\link{urlmaker_genemania}()},
\code{\link{urlmaker_hint}()},
\code{\link{urlmaker_hippie}()},
\code{\link{urlmaker_huri}()},
\code{\link{urlmaker_iid}()},
\code{\link{urlmaker_innatedb}()},
\code{\link{urlmaker_intact}()},
\code{\link{urlmaker_irefindex}()},
\code{\link{urlmaker_matrixdb}()},
\code{\link{urlmaker_mint}()},
\code{\link{urlmaker_pathwaycommons}()},
\code{\link{urlmaker_reactome}()}
}
\concept{urlmakers}
