% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_utils.R
\name{add_info_from_dataframe_to_graph}
\alias{add_info_from_dataframe_to_graph}
\title{add_info_from_dataframe_to_graph()}
\usage{
add_info_from_dataframe_to_graph(anno_df, g, idtype_anno = "gene_symbol")
}
\arguments{
\item{anno_df}{dataframe of ppi data from corresponding database containing annotation (Ensembl, GeneSymbol, Entrez, Uniprot)}

\item{g}{graph object created with build_graph()}

\item{idtype_anno}{Character, specifies which identifier type is used in the
\code{anno_df}. Defaults to "gene_symbol"}
}
\value{
An igraph object
}
\description{
add_info_from_dataframe_to_graph()
}
