% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-4_db_iid.R
\name{annotation_iid}
\alias{annotation_iid}
\title{annotation_iid ()}
\usage{
annotation_iid(ppi_iid, species, version)
}
\arguments{
\item{ppi_iid}{variable defined by ppis_iid in get_networkdata_iid()}

\item{species}{from which species does the data come from}

\item{version}{version of the data files in iid}
}
\value{
ppis_iid_annotated
}
\description{
annotation_iid ()
}
\examples{
#\donttest{
# annotation_iid(ppi_iid,
#               species = "human",
#               version = "2021-05")
#}
}
