% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-1_db_stringdb.R
\name{annotation_stringdb}
\alias{annotation_stringdb}
\title{annotation_stringdb ()}
\usage{
annotation_stringdb(ppi_stringdb, species, version, create_ppi_anno_df = TRUE)
}
\arguments{
\item{ppi_stringdb}{variable defined by ppis_stringdb in get_networkdata_stringdb()}

\item{species}{from which species does the data come from}

\item{version}{version of the data files in stringdb}

\item{create_ppi_anno_df}{TRUE=returning a dataframe that contains all annotations to interactor A & B in seperate columns, FALSE=returning the annotation datatframe (unique interactors (colnames = "uniprot_id", "gene_symbol", "entrez_id", "ensembl_id"))}
}
\value{
ppis_stringdb_annotated
}
\description{
annotation_stringdb ()
}
