% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_utils.R
\name{build_graph}
\alias{build_graph}
\title{build_graph()}
\usage{
build_graph(
  graph_data,
  data_source = c("stringdb", "hint", "funcoup", "iid", "irefindex", "mint", "genemania",
    "huri", "matrixdb", "pathwaycommons", "reactome", "innatedb", "biogrid", "intact"),
  output_format = "igraph",
  min_score_threshold = NULL,
  add_info_nodes = TRUE
)
}
\arguments{
\item{graph_data}{ppi data}

\item{data_source}{database}

\item{output_format}{selection of different graph functions that can be used}

\item{min_score_threshold}{select ppis that are "confident" depending on the scoretype/value}
}
\value{
my_graph
}
\description{
build_graph()
}
\examples{

db_matrixdb_df <- get_networkdata_matrixdb(
  species = "human",
  type = "CORE"
)

igraph_object_matrixdb <- build_graph(graph_data = db_matrixdb_df,
                                  data_source = "matrixdb",
                                  output_format = "igraph",
                                  min_score_threshold = NULL,
                                  add_info_nodes = TRUE)
igraph_object_matrixdb

}
