% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-3_db_funcoup.R
\name{build_graph_funcoup}
\alias{build_graph_funcoup}
\title{build_graph_funcoup()}
\usage{
build_graph_funcoup(
  graph_data,
  output_format = "igraph",
  min_score_threshold = NULL
)
}
\arguments{
\item{graph_data}{ppi data from funcoup}

\item{output_format}{selection of different graph functions that can be used}

\item{min_score_treshold}{select ppis that are "confident" depending on the scoretype/value}
}
\description{
build_graph_funcoup()
}
\examples{
\dontrun{

db_funcoup_df <- get_networkdata_funcoup(species = "H.sapiens",
                                         version = "5.0",
                                         type = "compact")

db_funcoup_graph <- build_graph_funcoup(graph_data = db_funcoup_df,
                                        output_format = "igraph",
                                        min_score_threshold = "0.2")
db_funcoup_graph #list of 17004
}


}
