% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_hippie_functions.R
\name{build_graph_hippie}
\alias{build_graph_hippie}
\title{build_graph_hippie()}
\usage{
build_graph_hippie(
  graph_data,
  output_format = "igraph",
  min_score_threshold = NULL
)
}
\arguments{
\item{graph_data}{ppi data from hippie}

\item{output_format}{selection of different graph functions that can be used}

\item{min_score_threshold}{select ppis that are "confident" depending on the scoretype/value}
}
\description{
build_graph_hippie()
}
\examples{
\dontrun{

db_hippie_df <- get_networkdata_hippie(species = "Homo_Sapiens",
                                       version = "current"
                                      )

db_hippie_graph <- build_graph_hippie(graph_data = db_hippie_df,
                                      output_format = "igraph",
                                      min_score_threshold = "0.35"
                                      )
db_hippie_graph
}


}
