% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-4_db_iid.R
\name{build_graph_iid}
\alias{build_graph_iid}
\title{build_graph_iid()}
\usage{
build_graph_iid(graph_data, output_format = "igraph")
}
\arguments{
\item{graph_data}{ppi data from iid}

\item{output_format}{selection of different graph functions that can be used}
}
\description{
build_graph_iid()
}
\examples{
\dontrun{

db_iid_df <- get_networkdata_iid(species = "human",
                                 version = "2021-05"
                                 )

db_iid_graph <- build_graph_iid(graph_data = db_iid_df,
                                output_format = "igraph")
db_iid_graph #list of
}


}
