% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-1_db_stringdb.R
\name{build_graph_stringdb}
\alias{build_graph_stringdb}
\title{build_graph_stringdb()}
\usage{
build_graph_stringdb(
  graph_data,
  output_format = "igraph",
  min_score_threshold = NULL
)
}
\arguments{
\item{graph_data}{ppi data from stringdb}

\item{output_format}{selection of different graph functions that can be used}

\item{min_score_threshold}{select ppis that are "confident" depending on the scoretype/value}
}
\description{
build_graph_stringdb()
}
\examples{
\dontrun{

db_stringdb_df <- get_networkdata_stringdb(species = "Mus musculus",
                                           version = "12.0",
                                           cache = TRUE,
                                           get_annotation = TRUE,
                                           add_annotation = TRUE
                                          )

db_stringdb_igraph_object <- build_graph_stringdb(graph_data = db_stringdb_df,
                                                  output_format = "igraph",
                                                  min_score_threshold = "0.35")
db_stringdb_graph #list of 17490
}
}
