% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_utils.R
\name{cache_NetworkHub}
\alias{cache_NetworkHub}
\title{Title}
\usage{
cache_NetworkHub(rname, fpath, nh_cachedir = "NetworkHub", ...)
}
\arguments{
\item{rname}{TODO}

\item{fpath}{TODO}

\item{nh_cachedir}{TODO}

\item{...}{TODO}
}
\value{
TODO
}
\description{
Title
}
\examples{
# for example, retrieve something from stringDB

url_sdb <- urlmaker_stringdb("PPI", "Homo sapiens", "12.0")
url_sdb

cache_NetworkHub(
  rname = "STRINGDB_Homo sapiens_v12.0",
  fpath = url_sdb
)
}
