% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cache_NetworkHub}
\alias{cache_NetworkHub}
\title{cache_NetworkHub()}
\usage{
cache_NetworkHub(rname, fpath, nh_cachedir = "NetworkHub", ...)
}
\arguments{
\item{rname}{resource name in the cache (local)}

\item{fpath}{filepath (external)}

\item{nh_cachedir}{default directory name where the cache will be stored}

\item{...}{further arguments passed to or from other methods}
}
\value{
rpath returns the resource path of the cached file
}
\description{
cache_NetworkHub()
}
\examples{
# for example, retrieve something from stringDB

url_stringdb <- urlmaker_stringdb(type = "PPI",
                                  species = "Homo sapiens",
                                  version = "12.0")
cache_NetworkHub(
  rname = "STRINGDB_Homo sapiens_v12.0",
  fpath = url_stringdb
)
}
