% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fetch_NetworkHub}
\alias{fetch_NetworkHub}
\title{fetch_NetworkHub()}
\usage{
fetch_NetworkHub(rname, update = TRUE, nh_cachedir = "NetworkHub", ...)
}
\arguments{
\item{rname}{resource name in the cache (local)}

\item{update}{default parameter set to TRUE (up to date version of the resource)}

\item{nh_cachedir}{default directory name where the cache will be stored}

\item{...}{further arguments passed to or from other methods}
}
\value{
res_nh returns rpath to res_nh or NULL if no entry
}
\description{
fetch_NetworkHub()
}
\examples{
fetch_NetworkHub(rname = "STRINGDB_Homo sapiens_v12.0")

}
