% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_utils.R
\name{fetch_NetworkHub}
\alias{fetch_NetworkHub}
\title{Title}
\usage{
fetch_NetworkHub(rname, update = TRUE, nh_cachedir = "NetworkHub", ...)
}
\arguments{
\item{rname}{TODO}

\item{update}{TODO}

\item{nh_cachedir}{TODO}

\item{...}{TODO}
}
\value{
TODO
}
\description{
Title
}
\examples{
fetch_NetworkHub(rname = "STRINGDB_Homo sapiens_v12.0")

}
