% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-1_db_stringdb.R
\name{get_accessoryinfo_stringdb}
\alias{get_accessoryinfo_stringdb}
\title{Get accessory info from STRINGDB}
\usage{
get_accessoryinfo_stringdb(species, version, cache = TRUE)
}
\arguments{
\item{species}{from which species does the data come from}

\item{version}{version of the data files in stringdb}

\item{cache}{default value set to TRUE (automatically checks if the data file is already stored in the cache)}
}
\value{
proteininfo_stringdb variable assigned to the datafile that contains info about protein annotation
}
\description{
Get accessory info from STRINGDB
}
\examples{
#\donttest{
#human_acc_stringdb <- get_accessoryinfo_stringdb(species = "Homo sapiens",
#                                       version = "12.0")
#}
}
