% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-1_db_stringdb.R
\name{get_annotation_stringdb}
\alias{get_annotation_stringdb}
\title{get_annotation_stringdb ()}
\usage{
get_annotation_stringdb(ppi_stringdb, species)
}
\arguments{
\item{ppi_stringdb}{variable defined by ppis_stringdb in get_networkdata_stringdb()}

\item{species}{from which species does the data come from}
}
\value{
anno_df (for corresponding species in stringdb)
}
\description{
get_annotation_stringdb ()
}
\examples{
\dontrun{

db_stringdb_df <- get_networkdata_stringdb(species = "Mus musculus",
                                           version = "12.0",
                                           cache = TRUE,
                                           get_annotation = FALSE,
                                           add_annotation = FALSE
                                          )

db_stringdb_anno_df <- get_annotation_stringdb(ppi_stringdb = db_stringdb_df,
                                                species = "Mus musculus"
                                              )
}
}
