% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-3_db_funcoup.R
\name{get_networkdata_funcoup}
\alias{get_networkdata_funcoup}
\title{get_networkdata_funcoup()}
\usage{
get_networkdata_funcoup(
  species = "H.sapiens",
  version = "5.0",
  cache = TRUE,
  add_annotation = TRUE,
  ...
)
}
\arguments{
\item{species}{from which species does the data come from c( "A.thaliana", "B.subtilis", "B.taurus", "C.elegans","C.familiaris", "C.intestinalis", "D.melanogatser", "D.rerio", "E.coli", "G.gallus", "H.sapiens", "M.jannaschii", "M.musculus", "O.sativa", "P.falciparum", "R.norvegicus", "S.cerevisae", "S.pombe", "S.scrofa", "S.solfataricus")}

\item{version}{version of the data files in funcoup}

\item{cache}{default value set to TRUE (automatically checks if the data file is already stored in the cache)}

\item{add_annotation}{expanding the dataframe with four columns (Entrez_ID and Ensembl_ID)}

\item{...}{further arguments passed to or from other methods}
}
\value{
ppis_funcoup

ppis_annotated_funcoup
}
\description{
get_networkdata_funcoup()
}
\examples{

db_funcoup_df <- get_networkdata_funcoup(species = "H.sapiens", version = "5.0")
db_funcoup_df

}
