% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_genemania_functions.R
\name{get_networkdata_genemania}
\alias{get_networkdata_genemania}
\title{get_networkdata_genemania()}
\usage{
get_networkdata_genemania(
  species = "Homo_sapiens",
  version = "current",
  cache = TRUE,
  get_annotation = TRUE,
  add_annotation = TRUE,
  ...
)
}
\arguments{
\item{species}{from which species does the data come from}

\item{version}{version of the data files in genemania}

\item{cache}{default value set to TRUE (automatically checks if the data file is already stored in the cache)}

\item{get_annotation}{creation of an annotation dataframe using AnnotationDbi packages, default value set to TRUE}

\item{add_annotation}{adding annotation to ppi dataframe, default value set to TRUE}

\item{...}{further arguments passed to or from other methods}
}
\value{
ppis_genemania

db_genemania_ppi_anno_df

db_genemania_anno_df
}
\description{
get_networkdata_genemania()
}
\examples{
\dontrun{
db_genemania_df <- get_networkdata_genemania(species = "Homo_sapiens",
                                             version = "current",
                                             get_annotation = FALSE,
                                             add_annotation = FALSE
                                             )
}
}
