% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-4_db_iid.R
\name{get_networkdata_iid}
\alias{get_networkdata_iid}
\title{get_networkdata_iid()}
\usage{
get_networkdata_iid(
  species,
  version = "2021-05",
  cache = TRUE,
  add_annotation = TRUE,
  ...
)
}
\arguments{
\item{species}{from which species does the data come from}

\item{version}{version of the data files in iid}

\item{cache}{default value set to TRUE (automatically checks if the data file is already stored in the cache)}

\item{add_annotation}{expanding the dataframe with four columns (Entrez_ID and Ensembl_ID)}

\item{...}{further arguments passed to or from other methods}
}
\value{
ppis_iid
}
\description{
get_networkdata_iid()
}
\examples{

db_iid_df <- get_networkdata_iid(
  species = "human",
  version = "2021-05"
)

db_iid_df

}
