% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_innatedb_functions.R
\name{get_networkdata_innatedb}
\alias{get_networkdata_innatedb}
\title{get_networkdata_innatedb()}
\usage{
get_networkdata_innatedb(
  species = "taxid:9606(Human)",
  version = "5.4",
  cache = TRUE,
  get_annotation = TRUE,
  add_annotation = TRUE,
  ...
)
}
\arguments{
\item{species}{default value = "taxid:9606(Homo sapiens)" - from which species does the data come from}

\item{version}{version of the data files in innatedb}

\item{cache}{default value set to TRUE (automatically checks if the data file is already stored in the cache)}

\item{get_annotation}{creation of an annotation dataframe using AnnotationDbi packages, default value set to TRUE}

\item{add_annotation}{adding annotation to ppi dataframe, default value set to TRUE}

\item{...}{further arguments passed to or from other methods}
}
\value{
ppi_innatedb
}
\description{
get_networkdata_innatedb()
}
\examples{
\dontrun{
db_innatedb_df <- get_networkdata_innatedb( species = "taxid:9606(Human)",
                                            version = "5.4",
                                            cache = TRUE,
                                            get_annotation = FALSE,
                                            add_annotation = FALSE
                                            )
}
}
