% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_intact_functions.R
\name{get_networkdata_intact}
\alias{get_networkdata_intact}
\title{get_networkdata_intact()}
\usage{
get_networkdata_intact(
  species = "taxid:9606(human)|taxid:9606(Homo sapiens)",
  version = "current",
  cache = TRUE,
  get_annotation = TRUE,
  add_annotation = TRUE,
  ...
)
}
\arguments{
\item{species}{default value = "taxid:9606(human)|taxid:9606(Homo sapiens)" - from which species does the data come from}

\item{version}{default value = "current" version of the data files in intact}

\item{cache}{default value set to TRUE (automatically checks if the data file is already stored in the cache)}

\item{get_annotation}{creation of an annotation dataframe using AnnotationDbi packages, default value set to TRUE}

\item{add_annotation}{adding annotation to ppi dataframe, default value set to TRUE}

\item{...}{further arguments passed to or from other methods}
}
\value{
ppis_intact

db_intact_ppi_anno_df

db_intact_anno_df
}
\description{
get_networkdata_intact()
}
\examples{
\dontrun{
db_intact_df <- get_networkdata_intact(
  species = "taxid:9606(human)|taxid:9606(Homo sapiens)",
  version = "current",
  cache = TRUE,
  get_annotation = FALSE,
  add_annotation = FALSE
)
}
}
