% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_irefindex_functions.R
\name{get_networkdata_irefindex}
\alias{get_networkdata_irefindex}
\title{get_networkdata_irefindex()}
\usage{
get_networkdata_irefindex(
  species = "Homo sapiens",
  version = "08-28-2023",
  cache = TRUE,
  get_annotation = TRUE,
  add_annotation = TRUE,
  ...
)
}
\arguments{
\item{species}{from which species does the data come from}

\item{version}{version of the data files in irefindex}

\item{cache}{default value set to TRUE (automatically checks if the data file is already stored in the cache)}

\item{get_annotation}{creation of an annotation dataframe using AnnotationDbi packages, default value set to TRUE}

\item{add_annotation}{adding annotation to ppi dataframe, default value set to TRUE}

\item{...}{further arguments passed to or from other methods}
}
\value{
ppi_irefindex
}
\description{
get_networkdata_irefindex()
}
\examples{
\dontrun{
db_irefindex_df <- get_networkdata_irefindex(species = "Homo sapiens",
                                             version = "08-28-2023",
                                             get_annotation = FALSE,
                                             add_annotation = FALSE
                                             )
db_irefindex_df
}
}
