% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_matrixdb_functions.R
\name{get_networkdata_matrixdb}
\alias{get_networkdata_matrixdb}
\title{get_networkdata_matrixdb()}
\usage{
get_networkdata_matrixdb(
  species = "human",
  type = "CORE",
  version = "4_0",
  cache = TRUE,
  get_annotation = TRUE,
  add_annotation = TRUE,
  ...
)
}
\arguments{
\item{species}{from which species does the data come from (default human because currently only human data provided from matrixdb)}

\item{type}{datasets provided by MatrixDB: "CORE" = MatrixDB manually curated interaction dataset}

\item{version}{version of the data files in MatrixDB}

\item{cache}{default value set to TRUE (automatically checks if the data file is already stored in the cache)}

\item{get_annotation}{creation of an annotation dataframe using AnnotationDbi packages, default value set to TRUE}

\item{add_annotation}{adding annotation to ppi dataframe, default value set to TRUE}

\item{...}{further arguments passed to or from other methods}
}
\value{
ppis_matrixdb

db_matrixdb_ppi_anno_df

db_matrixdb_anno_df
}
\description{
get_networkdata_matrixdb()
}
\examples{
\dontrun{
db_matrixdb_df <- get_networkdata_matrixdb( species = "human",
                                            type = "CORE",
                                            version = "4_0",
                                            cache = TRUE,
                                            get_annotation = FALSE,
                                            add_annotation = FALSE
                                          )

db_matrixdb_df
}
}
