% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_pathwaycommons_functions.R
\name{get_networkdata_pathwaycommons}
\alias{get_networkdata_pathwaycommons}
\title{get_networkdata_pathwaycommons()}
\usage{
get_networkdata_pathwaycommons(
  species = "human",
  version = "v12",
  cache = TRUE,
  get_annotation = TRUE,
  add_annotation = TRUE,
  ...
)
}
\arguments{
\item{species}{from which species does the data come from}

\item{version}{version of the data files in pathwaycommons}

\item{cache}{default value set to TRUE (automatically checks if the data file is already stored in the cache)}

\item{get_annotation}{creation of an annotation dataframe using AnnotationDbi packages, default value set to TRUE}

\item{add_annotation}{adding annotation to ppi dataframe, default value set to TRUE}

\item{...}{further arguments passed to or from other methods}
}
\value{
ppis_pathwaycommons

db_pathwaycommons_ppi_anno_df

db_pathwaycommons_anno_df
}
\description{
get_networkdata_pathwaycommons()
}
\examples{
\dontrun{
db_pathwaycommons_df <- get_networkdata_pathwaycommons( species = "human",
                                                        version = "v12",
                                                        cache = TRUE,
                                                        get_annotation = FALSE,
                                                        add_annotation = FALSE
                                                        )
}
}
