% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-1_db_stringdb.R
\name{get_networkdata_stringdb}
\alias{get_networkdata_stringdb}
\title{get_networkdata_stringdb()}
\usage{
get_networkdata_stringdb(
  species,
  version,
  cache = TRUE,
  get_annotation = TRUE,
  add_annotation = TRUE,
  ...
)
}
\arguments{
\item{species}{from which species does the data come from}

\item{version}{version of the data files in stringdb}

\item{cache}{default value set to TRUE (automatically checks if the data file is already stored in the cache)}

\item{get_annotation}{creation of an annotation dataframe , default value set to TRUE}

\item{add_annotation}{adding annotation to ppi dataframe, default value set to TRUE}

\item{...}{further arguments passed to or from other methods}
}
\value{
ppis_stringdb
}
\description{
get_networkdata_stringdb()
}
\examples{
\dontrun{
db_string_df <- get_networkdata_stringdb(species = "Homo sapiens",
                                         version = "12.0",
                                         cache = TRUE,
                                         get_annotation = FALSE,
                                         add_annotation = FALSE
                                        )
db_string_df
}

}
