% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{initialize_NetworkHub}
\alias{initialize_NetworkHub}
\title{initialize_NetworkHub()}
\usage{
initialize_NetworkHub(nh_cachedir = "NetworkHub")
}
\arguments{
\item{nh_cachedir}{default directory name where the cache will be stored}
}
\value{
bfc_nh variable assigned to the cache
}
\description{
initialize_NetworkHub()
}
\examples{
library("BiocFileCache") # load the library of the package

# use this function to create the cache called "bfc_nh"
bfc_nh <- initialize_NetworkHub(nh_cachedir = "NetworkHub")

# get the path to the cache
bfccache(bfc_nh)

# how many entries?
length(bfc_nh)
# an overview on the entries
bfcinfo(bfc_nh)
# bfcremove(bfc_nh, c("BFC10", "BFC11", "BFC12", "BFC13"))

}
