% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_utils.R
\name{initialize_NetworkHub}
\alias{initialize_NetworkHub}
\title{initialize_NetworkHub()}
\usage{
initialize_NetworkHub(nh_cachedir = "NetworkHub")
}
\arguments{
\item{nh_cachedir}{default directory name where the cache will be stored}

\item{cache_dir}{user directory of the cache}
}
\value{
bfc_nh variabe assigned to the chache
}
\description{
initialize_NetworkHub()
}
\examples{
library("BiocFileCache") # load the library of the package
bfc_nh <- initialize_NetworkHub() # use this function to create the cache called "bfc_nh"
bfccache(bfc_nh) # get the path to the cache
length(bfc_nh) # how many entries?
bfcinfo(bfc_nh) # what are the column names of the cache

# nh_cachedir = default directory name where the cache will be stored
}
