% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_NetworkHub-pkg.R
\docType{package}
\name{NetworkHub-pkg}
\alias{NetworkHub}
\alias{NetworkHub-package}
\alias{NetworkHub-pkg}
\title{NetworkHub}
\description{
NetworkHub
}
\details{
This package provides functionality to ... TODO
}
\author{
\strong{Maintainer}: Lotta wagnerlo@uni-mainz.de Wagner (\href{https://orcid.org/0009-0009-7545-0173}{ORCID})

Authors:
\itemize{
  \item federico marinif@uni-mainz.de Marini (\href{https://orcid.org/0000-0003-3252-7758}{ORCID})
}

}
\keyword{internal}
