% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkHub-pkg.R
\docType{package}
\name{NetworkHub-pkg}
\alias{NetworkHub}
\alias{NetworkHub-package}
\alias{NetworkHub-pkg}
\title{NetworkHub}
\description{
NetworkHub
}
\details{
This package provides functionality to ... TODO
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lottawagner/NetworkHub}
  \item Report bugs at \url{https://github.com/lottawagner/NetworkHub/issues}
}

}
\author{
\strong{Maintainer}: Lotta Wagner \email{wagnerlo@uni-mainz.de} (\href{https://orcid.org/0009-0009-7545-0173}{ORCID})

Authors:
\itemize{
  \item Federico Marini \email{marinif@uni-mainz.de} (\href{https://orcid.org/0000-0003-3252-7758}{ORCID})
}

}
\keyword{internal}
