% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urlmaker_functions.R
\name{urlmaker_biogrid}
\alias{urlmaker_biogrid}
\title{urlmaker_biogrid()}
\usage{
urlmaker_biogrid(version = "4.4.238")
}
\arguments{
\item{version}{version of the data files in BioGRID}
}
\value{
url returns the corresponding url set by params #SPECIESDEFINITION
}
\description{
urlmaker_biogrid()
}
\examples{
url_biogrid <- urlmaker_biogrid()
url_biogrid
}
\seealso{
Other urlmakers: 
\code{\link{urlmaker_cpdb}()},
\code{\link{urlmaker_funcoup}()},
\code{\link{urlmaker_genemania}()},
\code{\link{urlmaker_hint}()},
\code{\link{urlmaker_hippie}()},
\code{\link{urlmaker_huri}()},
\code{\link{urlmaker_iid}()},
\code{\link{urlmaker_innatedb}()},
\code{\link{urlmaker_intact}()},
\code{\link{urlmaker_irefindex}()},
\code{\link{urlmaker_matrixdb}()},
\code{\link{urlmaker_mint}()},
\code{\link{urlmaker_pathwaycommons}()},
\code{\link{urlmaker_reactome}()},
\code{\link{urlmaker_stringdb}()}
}
\concept{urlmakers}
