% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_urlmaker.R
\name{urlmaker_biogrid}
\alias{urlmaker_biogrid}
\title{info_species_biogrid}
\usage{
urlmaker_biogrid(type = "PPI", species = "Homo sapiens", version = "4.4.235")
}
\arguments{
\item{version}{version of the data files in stringdb}
}
\value{
url returns the corresponding url set by params
}
\description{
info_species_biogrid
}
\examples{
# TODO
urlmaker_biogrid
}
