% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_urlmaker.R
\name{urlmaker_biogrid}
\alias{urlmaker_biogrid}
\title{urlmaker_biogrid()}
\usage{
urlmaker_biogrid(version = "4.4.236")
}
\arguments{
\item{version}{version of the data files in BioGRID}
}
\value{
url returns the corresponding url set by params #SPECIESDEFINITION
}
\description{
urlmaker_biogrid()
}
\examples{
url_biogrid <- urlmaker_biogrid()
url_biogrid
}
