% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_urlmaker.R
\name{urlmaker_cpdb}
\alias{urlmaker_cpdb}
\title{urlmaker_cpdb()}
\usage{
urlmaker_cpdb(species = "human")
}
\arguments{
\item{species}{currently only human (default value)}
}
\value{
url returns the corresponding url set by params
}
\description{
urlmaker_cpdb()
}
\examples{
url_cpdb <- urlmaker_cpdb(species = "human")
url_cpdb
}
