% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urlmaker_functions.R
\name{urlmaker_funcoup}
\alias{urlmaker_funcoup}
\title{urlmaker_funcoup()}
\usage{
urlmaker_funcoup(
  version = "6.0",
  species = "H.sapiens",
  type = c("compact", "full")
)
}
\arguments{
\item{version}{version of FunCoup , default value = "5.0", value as type "5.0" #UPDATEVERSION}

\item{species}{types listed in list_species_funcoup depending on current version, default value = "H.sapiens" #UPDATEVERSION}

\item{type}{interaction types in FunCoup (for current version compact recommended, for older versions only full possible)}
}
\value{
url returns the corresponding url set by params
}
\description{
urlmaker_funcoup()
}
\examples{
url_funcoup <- urlmaker_funcoup(version = "6.0",
                                species = "B.taurus",
                                type = "full")
url_funcoup
}
