% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urlmaker_functions.R
\name{urlmaker_genemania}
\alias{urlmaker_genemania}
\title{urlmaker_genemania()}
\usage{
urlmaker_genemania(species = "Homo_sapiens", version = "current")
}
\arguments{
\item{species}{types listed in list_species_genemania depending on current version, default value = "Homo_sapiens"}

\item{version}{version of GeneMania , default value = "current" #UPDATEVERSION}
}
\value{
url returns the corresponding url set by params
}
\description{
urlmaker_genemania()
}
\examples{
url_genemania <- urlmaker_genemania(
  species = "Homo_sapiens",
  version = "current"
)
url_genemania
}
