% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urlmaker_functions.R
\name{urlmaker_genemania}
\alias{urlmaker_genemania}
\title{urlmaker_genemania()}
\usage{
urlmaker_genemania(species = "Homo_sapiens", version = "current")
}
\arguments{
\item{species}{types listed in list_species_genemania depending on current version, default value = "Homo_sapiens"}

\item{version}{version of GeneMania , default value = "current" #UPDATEVERSION}
}
\value{
url returns the corresponding url set by params
}
\description{
urlmaker_genemania()
}
\examples{
url_genemania <- urlmaker_genemania(
  species = "Homo_sapiens",
  version = "current"
)
url_genemania
}
\seealso{
Other urlmakers: 
\code{\link{urlmaker_biogrid}()},
\code{\link{urlmaker_cpdb}()},
\code{\link{urlmaker_funcoup}()},
\code{\link{urlmaker_hint}()},
\code{\link{urlmaker_hippie}()},
\code{\link{urlmaker_huri}()},
\code{\link{urlmaker_iid}()},
\code{\link{urlmaker_innatedb}()},
\code{\link{urlmaker_intact}()},
\code{\link{urlmaker_irefindex}()},
\code{\link{urlmaker_matrixdb}()},
\code{\link{urlmaker_mint}()},
\code{\link{urlmaker_pathwaycommons}()},
\code{\link{urlmaker_reactome}()},
\code{\link{urlmaker_stringdb}()}
}
\concept{urlmakers}
