% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urlmaker_functions.R
\name{urlmaker_hint}
\alias{urlmaker_hint}
\title{urlmaker_hint()}
\usage{
urlmaker_hint(type = "binary", species = "HomoSapiens", version)
}
\arguments{
\item{type}{interaction types in HINT, default value = "binary"}

\item{species}{types listed in list_species_hint depending on current version, default value = "HomoSapiens" #UPDATEVERSION}

\item{version}{version of HINT, always written as year - month ( 2024-06, 2020-08, ...)}
}
\value{
url returns the corresponding url set by params
}
\description{
urlmaker_hint()
}
\examples{
url_hint <- urlmaker_hint(
  type = "binary",
  species = "HomoSapiens",
  version = "2024-06"
)
url_hint
}
\seealso{
Other urlmakers: 
\code{\link{urlmaker_biogrid}()},
\code{\link{urlmaker_cpdb}()},
\code{\link{urlmaker_funcoup}()},
\code{\link{urlmaker_genemania}()},
\code{\link{urlmaker_hippie}()},
\code{\link{urlmaker_huri}()},
\code{\link{urlmaker_iid}()},
\code{\link{urlmaker_innatedb}()},
\code{\link{urlmaker_intact}()},
\code{\link{urlmaker_irefindex}()},
\code{\link{urlmaker_matrixdb}()},
\code{\link{urlmaker_mint}()},
\code{\link{urlmaker_pathwaycommons}()},
\code{\link{urlmaker_reactome}()},
\code{\link{urlmaker_stringdb}()}
}
\concept{urlmakers}
