% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urlmaker_functions.R
\name{urlmaker_hippie}
\alias{urlmaker_hippie}
\title{urlmaker_hippie()}
\usage{
urlmaker_hippie(species = "Homo_sapiens", version = "current")
}
\arguments{
\item{species}{default value = "Homo_sapiens", because this database only provides human data}

\item{version}{default value = "current", version of the database ... #UPDATEVERSION}
}
\value{
url returns the corresponding url set by params
}
\description{
urlmaker_hippie()
}
\examples{

url_hippie <- urlmaker_hippie(
  species = "Homo_sapiens",
  version = "current"
)
url_hippie

}
\seealso{
Other urlmakers: 
\code{\link{urlmaker_biogrid}()},
\code{\link{urlmaker_cpdb}()},
\code{\link{urlmaker_funcoup}()},
\code{\link{urlmaker_genemania}()},
\code{\link{urlmaker_hint}()},
\code{\link{urlmaker_huri}()},
\code{\link{urlmaker_iid}()},
\code{\link{urlmaker_innatedb}()},
\code{\link{urlmaker_intact}()},
\code{\link{urlmaker_irefindex}()},
\code{\link{urlmaker_matrixdb}()},
\code{\link{urlmaker_mint}()},
\code{\link{urlmaker_pathwaycommons}()},
\code{\link{urlmaker_reactome}()},
\code{\link{urlmaker_stringdb}()}
}
\concept{urlmakers}
