% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urlmaker_functions.R
\name{urlmaker_huri}
\alias{urlmaker_huri}
\title{urlmaker_huri()}
\usage{
urlmaker_huri(species = "human", type = c("HI-union", "Lit-BM"))
}
\arguments{
\item{species}{default value = "human", because this database only provides human data}

\item{type}{different datasets , more information on "http://www.interactome-atlas.org/about/"}
}
\value{
url returns the corresponding url set by params
}
\description{
urlmaker_huri()
}
\examples{

url_huri <- urlmaker_huri(
  species = "human",
  type = "HI-union"
)
url_huri
}
\seealso{
Other urlmakers: 
\code{\link{urlmaker_biogrid}()},
\code{\link{urlmaker_cpdb}()},
\code{\link{urlmaker_funcoup}()},
\code{\link{urlmaker_genemania}()},
\code{\link{urlmaker_hint}()},
\code{\link{urlmaker_hippie}()},
\code{\link{urlmaker_iid}()},
\code{\link{urlmaker_innatedb}()},
\code{\link{urlmaker_intact}()},
\code{\link{urlmaker_irefindex}()},
\code{\link{urlmaker_matrixdb}()},
\code{\link{urlmaker_mint}()},
\code{\link{urlmaker_pathwaycommons}()},
\code{\link{urlmaker_reactome}()},
\code{\link{urlmaker_stringdb}()}
}
\concept{urlmakers}
