% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_urlmaker.R
\name{urlmaker_huri}
\alias{urlmaker_huri}
\title{urlmaker_huri}
\usage{
urlmaker_huri(species = "human", type = "HI-union")
}
\arguments{
\item{species}{default value = "human", because this database only provides human data}

\item{type}{different datasets , more information on "http://www.interactome-atlas.org/about/"}
}
\value{
url returns the corresponding url set by params
}
\description{
urlmaker_huri
}
\examples{

urlmaker_huri(species = "human",
             type = "HI-union")
}
