% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urlmaker_functions.R
\name{urlmaker_iid}
\alias{urlmaker_iid}
\title{urlmaker_iid()}
\usage{
urlmaker_iid(species = "human", version = "2021-05")
}
\arguments{
\item{species}{types listed in list_species_iid depending on current version, default value = "human" #UPDATEVERSION}

\item{version}{version of IID , default value = "2021-05", value as type "2021-05" #UPDATEVERSION}
}
\value{
url returns the corresponding url set by params
}
\description{
urlmaker_iid()
}
\examples{
url_iid <- urlmaker_iid(species = "mouse")
url_iid
}
\seealso{
Other urlmakers: 
\code{\link{urlmaker_biogrid}()},
\code{\link{urlmaker_cpdb}()},
\code{\link{urlmaker_funcoup}()},
\code{\link{urlmaker_genemania}()},
\code{\link{urlmaker_hint}()},
\code{\link{urlmaker_hippie}()},
\code{\link{urlmaker_huri}()},
\code{\link{urlmaker_innatedb}()},
\code{\link{urlmaker_intact}()},
\code{\link{urlmaker_irefindex}()},
\code{\link{urlmaker_matrixdb}()},
\code{\link{urlmaker_mint}()},
\code{\link{urlmaker_pathwaycommons}()},
\code{\link{urlmaker_reactome}()},
\code{\link{urlmaker_stringdb}()}
}
\concept{urlmakers}
