% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_urlmaker.R
\name{urlmaker_iid}
\alias{urlmaker_iid}
\title{urlmaker_iid()}
\usage{
urlmaker_iid(species = "human", version = "2021-05")
}
\arguments{
\item{species}{types listed in list_species_iid depending on current version, default value = "human" #UPDATEVERSION}

\item{version}{version of IID , default value = "2021-05", value as type "2021-05" #UPDATEVERSION}
}
\value{
url returns the corresponding url set by params
}
\description{
urlmaker_iid()
}
\examples{
url_iid <- urlmaker_iid(species = "mouse")
url_iid
}
