% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urlmaker_functions.R
\name{urlmaker_innatedb}
\alias{urlmaker_innatedb}
\title{urlmaker_innatedb()}
\usage{
urlmaker_innatedb(url = url, version = "5.4")
}
\arguments{
\item{url}{innatedb doesn't provide information about the version or the species in their url, that why the default value of url is url}

\item{version}{parameter set to current version 5.4 #UPDATEVERSION}
}
\value{
url
}
\description{
urlmaker_innatedb()
}
\examples{
url_innatedb <- urlmaker_innatedb()
url_innatedb
}
