% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_urlmaker.R
\name{urlmaker_intact}
\alias{urlmaker_intact}
\title{urlmaker_intact()}
\usage{
urlmaker_intact(version = "current")
}
\arguments{
\item{version}{version of the data files in IntAct, default value = "current" (August 2024  = 2024-05-23 18:09    6.6G)}
}
\value{
url returns the corresponding url set by params #SPECIESDEFINITION later on
}
\description{
urlmaker_intact()
}
\examples{
urlmaker_intact(version = "current")
}
