% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urlmaker_functions.R
\name{urlmaker_irefindex}
\alias{urlmaker_irefindex}
\title{#' urlmaker_irefindex()}
\usage{
urlmaker_irefindex(species, version = "08-28-2023")
}
\arguments{
\item{species}{from which species does the data come from, default value = "Homo sapiens"}

\item{version}{version of data files in iRefIndex, default value = "2023-08-29" #UPDATEVERSION}
}
\value{
url returns the corresponding url set by params
}
\description{
#' urlmaker_irefindex()
}
\examples{
url_irefindex <- urlmaker_irefindex(
  species = "Homo sapiens",
  version = "08-28-2023"
)
url_irefindex
}
\seealso{
Other urlmakers: 
\code{\link{urlmaker_biogrid}()},
\code{\link{urlmaker_cpdb}()},
\code{\link{urlmaker_funcoup}()},
\code{\link{urlmaker_genemania}()},
\code{\link{urlmaker_hint}()},
\code{\link{urlmaker_hippie}()},
\code{\link{urlmaker_huri}()},
\code{\link{urlmaker_iid}()},
\code{\link{urlmaker_innatedb}()},
\code{\link{urlmaker_intact}()},
\code{\link{urlmaker_matrixdb}()},
\code{\link{urlmaker_mint}()},
\code{\link{urlmaker_pathwaycommons}()},
\code{\link{urlmaker_reactome}()},
\code{\link{urlmaker_stringdb}()}
}
\concept{urlmakers}
