% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_urlmaker.R
\name{urlmaker_irefindex}
\alias{urlmaker_irefindex}
\title{#' urlmaker_irefindex()}
\usage{
urlmaker_irefindex(species, version = "08-28-2023")
}
\arguments{
\item{species}{from which species does the data come from, default value = "Homo sapiens"}

\item{version}{version of data files in iRefIndex, default value = "2023-08-29" #UPDATEVERSION}
}
\value{
url returns the corresponding url set by params
}
\description{
#' urlmaker_irefindex()
}
\examples{
urlmaker_irefindex <- function(species = "Homo sapiens",
                                version = "08-28-2023")

}
