% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urlmaker_functions.R
\name{urlmaker_matrixdb}
\alias{urlmaker_matrixdb}
\title{urlmaker_matrixdb()}
\usage{
urlmaker_matrixdb(species = "human", type = c("all", "CORE"), version = "4_0")
}
\arguments{
\item{species}{default value = "human", because only one version and one species at MatrixDB #UPDATEVERSION}

\item{type}{datasets provided by MatrixDB: "CORE" = MatrixDB manually curated interaction dataset}

\item{version}{TODO, 4.0 was recently added to the URL}
}
\value{
url returns the corresponding url set by params
}
\description{
urlmaker_matrixdb()
}
\examples{
url_matrixdb <- urlmaker_matrixdb(type = "CORE")

url_matrixdb

# https://matrixdb.univ-lyon1.fr/downloads/matrixdb_CORE_4_0.tab.zip
# https://matrixdb.univ-lyon1.fr/downloads/matrixdb_all_4_0.tab.zip
}
\seealso{
Other urlmakers: 
\code{\link{urlmaker_biogrid}()},
\code{\link{urlmaker_cpdb}()},
\code{\link{urlmaker_funcoup}()},
\code{\link{urlmaker_genemania}()},
\code{\link{urlmaker_hint}()},
\code{\link{urlmaker_hippie}()},
\code{\link{urlmaker_huri}()},
\code{\link{urlmaker_iid}()},
\code{\link{urlmaker_innatedb}()},
\code{\link{urlmaker_intact}()},
\code{\link{urlmaker_irefindex}()},
\code{\link{urlmaker_mint}()},
\code{\link{urlmaker_pathwaycommons}()},
\code{\link{urlmaker_reactome}()},
\code{\link{urlmaker_stringdb}()}
}
\concept{urlmakers}
