% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_urlmaker.R
\name{urlmaker_matrixdb}
\alias{urlmaker_matrixdb}
\title{urlmaker_matrixdb()}
\usage{
urlmaker_matrixdb(species = "human")
}
\arguments{
\item{species}{default value = "human", because only one version and one species at MatrixDB #UPDATEVERSION}
}
\value{
url returns the corresponding url set by params
}
\description{
urlmaker_matrixdb()
}
\examples{
url_matrixdb <- urlmaker_matrixdb()
url_matrixdb
}
