% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urlmaker_functions.R
\name{urlmaker_mint}
\alias{urlmaker_mint}
\title{#' urlmaker_mint()}
\usage{
urlmaker_mint(species = "Homo Sapiens", version = "current")
}
\arguments{
\item{species}{from which species does the data come from, default value = "Homo sapiens"}

\item{version}{version of data files in MINT, default value = "current" #CURRENTVERSION}
}
\value{
url returns the corresponding url set by params
}
\description{
#' urlmaker_mint()
}
\examples{
url_mint <- urlmaker_mint(species = "Homo Sapiens")
url_mint
}
\seealso{
Other urlmakers: 
\code{\link{urlmaker_biogrid}()},
\code{\link{urlmaker_cpdb}()},
\code{\link{urlmaker_funcoup}()},
\code{\link{urlmaker_genemania}()},
\code{\link{urlmaker_hint}()},
\code{\link{urlmaker_hippie}()},
\code{\link{urlmaker_huri}()},
\code{\link{urlmaker_iid}()},
\code{\link{urlmaker_innatedb}()},
\code{\link{urlmaker_intact}()},
\code{\link{urlmaker_irefindex}()},
\code{\link{urlmaker_matrixdb}()},
\code{\link{urlmaker_pathwaycommons}()},
\code{\link{urlmaker_reactome}()},
\code{\link{urlmaker_stringdb}()}
}
\concept{urlmakers}
