% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_urlmaker.R
\name{urlmaker_mint}
\alias{urlmaker_mint}
\title{#' urlmaker_mint()}
\usage{
urlmaker_mint(species = "Homo Sapiens", version = "current")
}
\arguments{
\item{species}{from which species does the data come from, default value = "Homo sapiens"}

\item{version}{version of data files in MINT, default value = "current" #CURRENTVERSION}
}
\value{
url returns the corresponding url set by params
}
\description{
#' urlmaker_mint()
}
\examples{
url_mint <- urlmaker_mint(species = "Homo Sapiens")
url_mint
}
