% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urlmaker_functions.R
\name{urlmaker_pathwaycommons}
\alias{urlmaker_pathwaycommons}
\title{urlmaker_pathwaycommons()}
\usage{
urlmaker_pathwaycommons(species = "human", version = "v12")
}
\arguments{
\item{species}{default value = "Homo sapiens", because only one species}

\item{version}{default value = "v12" because v14 doesn't contain all datafiles #UPDATEVERSION}
}
\value{
url returns the corresponding url set by params
}
\description{
urlmaker_pathwaycommons()
}
\examples{

url_pc <- urlmaker_pathwaycommons()
url_pc
}
\seealso{
Other urlmakers: 
\code{\link{urlmaker_biogrid}()},
\code{\link{urlmaker_cpdb}()},
\code{\link{urlmaker_funcoup}()},
\code{\link{urlmaker_genemania}()},
\code{\link{urlmaker_hint}()},
\code{\link{urlmaker_hippie}()},
\code{\link{urlmaker_huri}()},
\code{\link{urlmaker_iid}()},
\code{\link{urlmaker_innatedb}()},
\code{\link{urlmaker_intact}()},
\code{\link{urlmaker_irefindex}()},
\code{\link{urlmaker_matrixdb}()},
\code{\link{urlmaker_mint}()},
\code{\link{urlmaker_reactome}()},
\code{\link{urlmaker_stringdb}()}
}
\concept{urlmakers}
