% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_urlmaker.R
\name{urlmaker_pathwaycommons}
\alias{urlmaker_pathwaycommons}
\title{urlmaker_pathwaycommons()}
\usage{
urlmaker_pathwaycommons(species = "human", version = "v12")
}
\arguments{
\item{species}{default value = "Homo sapiens", because only one species}

\item{version}{default value = "v12" because v14 doesn't contain all datafiles #UPDATEVERSION}
}
\value{
url returns the corresponding url set by params
}
\description{
urlmaker_pathwaycommons()
}
\examples{

url_pc <- urlmaker_pathwaycommons()
url_pc
}
