% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_urlmaker.R
\name{urlmaker_reactome}
\alias{urlmaker_reactome}
\title{urlmaker_reactome()}
\usage{
urlmaker_reactome(version = "current")
}
\arguments{
\item{version}{default value = "current" #UPDATEVERSION}
}
\value{
url returns the corresponding url set by params
}
\description{
urlmaker_reactome()
}
\examples{
urlmaker_reactome()
}
