% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urlmaker_functions.R
\name{urlmaker_stringdb}
\alias{urlmaker_stringdb}
\title{urlmaker_stringdb()}
\usage{
urlmaker_stringdb(
  type = c("PPI", "protein_info"),
  species = "Homo sapiens",
  version = "12.0"
)
}
\arguments{
\item{type}{which data file do you want to download ? ("PPI", "protein_info")}

\item{species}{from which species does the data come from}

\item{version}{version of the data files in stringdb}
}
\value{
url returns the corresponding url set by params
}
\description{
urlmaker_stringdb()
}
\examples{
url_stringdb <- urlmaker_stringdb(
  type = "PPI",
  species = "Homo sapiens",
  version = "12.0"
)

url_stringdb
}
\seealso{
Other urlmakers: 
\code{\link{urlmaker_biogrid}()},
\code{\link{urlmaker_cpdb}()},
\code{\link{urlmaker_funcoup}()},
\code{\link{urlmaker_genemania}()},
\code{\link{urlmaker_hint}()},
\code{\link{urlmaker_hippie}()},
\code{\link{urlmaker_huri}()},
\code{\link{urlmaker_iid}()},
\code{\link{urlmaker_innatedb}()},
\code{\link{urlmaker_intact}()},
\code{\link{urlmaker_irefindex}()},
\code{\link{urlmaker_matrixdb}()},
\code{\link{urlmaker_mint}()},
\code{\link{urlmaker_pathwaycommons}()},
\code{\link{urlmaker_reactome}()}
}
\concept{urlmakers}
