% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_urlmaker.R
\name{urlmaker_stringdb}
\alias{urlmaker_stringdb}
\title{urlmaker_stringdb()}
\usage{
urlmaker_stringdb(type = "PPI", species = "Homo sapiens", version = "12.0")
}
\arguments{
\item{type}{type of information that is stored in the file}

\item{species}{from which species does the data come from}

\item{version}{version of the data files in stringdb}
}
\value{
url returns the corresponding url set by params
}
\description{
urlmaker_stringdb()
}
\examples{
url_stringdb <- urlmaker_stringdb(type = "PPI",
                                  species = "Homo sapiens",
                                  version = "12.0")
url_stringdb
}
